refactor: split SportsGamesCenter into focused sub-components
