+++ src/app/[locale]/(platform)/sports/_components/SportsGamesCenter.tsx
@@ -2,56 +2,29 @@
 
 import type { Route } from 'next'
 import type {
-  CSSProperties,
   KeyboardEvent as ReactKeyboardEvent,
   MouseEvent as ReactMouseEvent,
-  MouseEvent as ReactMouseEventType,
 } from 'react'
-import type { EventOrderPanelOutcomeSelectedAccent }
-  from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelOutcomeButton'
-import type { SportsGamesButton, SportsGamesCard } from '@/app/[locale]/(platform)/sports/_utils/sports-games-data'
-import type { OddsFormat } from '@/lib/odds-format'
-import type { NormalizedBookLevel } from '@/lib/order-panel-utils'
-import type { SportsVertical } from '@/lib/sports-vertical'
-import type { Event, Market, Outcome, UserPosition } from '@/types'
-import type { DataPoint, PredictionChartCursorSnapshot, PredictionChartProps } from '@/types/PredictionChartTypes'
-import { useQuery } from '@tanstack/react-query'
+import type { SportsGamesCenterProps, SportsGamesMarketType } from './_sports-games-center/sports-games-center-types'
+import type { SportsGamesCard } from '@/app/[locale]/(platform)/sports/_utils/sports-games-data'
 import {
   CheckIcon,
-  ChevronDownIcon,
-  ChevronLeftIcon,
   ChevronRightIcon,
-  EqualIcon,
   RadioIcon,
-  RefreshCwIcon,
   SearchIcon,
   SettingsIcon,
   XIcon,
 } from 'lucide-react'
-import { useExtracted, useLocale } from 'next-intl'
-import dynamic from 'next/dynamic'
+import { useLocale } from 'next-intl'
 import Image from 'next/image'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import SellPositionModal from '@/app/[locale]/(platform)/_components/SellPositionModal'
-import EventChartControls from '@/app/[locale]/(platform)/event/[slug]/_components/EventChartControls'
-import EventChartEmbedDialog from '@/app/[locale]/(platform)/event/[slug]/_components/EventChartEmbedDialog'
-import EventChartExportDialog from '@/app/[locale]/(platform)/event/[slug]/_components/EventChartExportDialog'
-import EventConvertPositionsDialog from '@/app/[locale]/(platform)/event/[slug]/_components/EventConvertPositionsDialog'
-import { useOptionalMarketChannelSubscription } from '@/app/[locale]/(platform)/event/[slug]/_components/EventMarketChannelProvider'
-import EventOrderBook, { useOrderBookSummaries } from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderBook'
 import EventOrderPanelForm from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelForm'
 import EventOrderPanelMobile from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelMobile'
 import EventOrderPanelTermsDisclaimer
   from '@/app/[locale]/(platform)/event/[slug]/_components/EventOrderPanelTermsDisclaimer'
-import { TIME_RANGES, useEventPriceHistory } from '@/app/[locale]/(platform)/event/[slug]/_hooks/useEventPriceHistory'
-import { loadStoredChartSettings, storeChartSettings } from '@/app/[locale]/(platform)/event/[slug]/_utils/chartSettingsStorage'
-import { fetchOrderBookSummaries } from '@/app/[locale]/(platform)/event/[slug]/_utils/EventOrderBookUtils'
-import SportsEventAboutPanel from '@/app/[locale]/(platform)/sports/_components/SportsEventAboutPanel'
 import SportsLivestreamFloatingPlayer
   from '@/app/[locale]/(platform)/sports/_components/SportsLivestreamFloatingPlayer'
 import {
   hasSportsGamesCardPrimaryMarketTrio,
-  isSportsGamesCardResolved,
   resolveSportsGamesCardVisibleMarketTypes,
   resolveSportsGamesHeaderMarketTypes,
 } from '@/app/[locale]/(platform)/sports/_utils/sports-games-data'
@@ -69,4991 +42,69 @@ import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 import { useCurrentTimestamp } from '@/hooks/useCurrentTimestamp'
 import { useIsMobile } from '@/hooks/useIsMobile'
-import { useWindowSize } from '@/hooks/useWindowSize'
 import { useRouter } from '@/i18n/navigation'
-import { MICRO_UNIT, ORDER_SIDE, ORDER_TYPE, OUTCOME_INDEX } from '@/lib/constants'
-import { fetchUserPositionsForMarket } from '@/lib/data-api/user'
-import {
-  formatAmountInputValue,
-  formatCentsLabel,
-  formatCurrency,
-  formatSharePriceLabel,
-  formatSharesLabel,
-  formatVolume,
-  fromMicro,
-} from '@/lib/formatters'
-import { resolveOutcomePriceCents, resolveOutcomeSelectionPriceCents } from '@/lib/market-pricing'
-import { formatOddsFromCents, ODDS_FORMAT_OPTIONS } from '@/lib/odds-format'
-import { calculateMarketFill, normalizeBookLevels } from '@/lib/order-panel-utils'
-import { calculateYAxisBounds } from '@/lib/prediction-chart'
-import {
-  isSportsTeamTone,
-  resolveSportsTeamFallbackButtonStyle,
-  resolveSportsTeamFallbackColor,
-  resolveSportsTeamFallbackDepthStyle,
-  resolveSportsTeamFallbackOverlayStyle,
-} from '@/lib/sports-team-colors'
+import { formatVolume } from '@/lib/formatters'
+import { ODDS_FORMAT_OPTIONS } from '@/lib/odds-format'
 import { shouldUseCroppedSportsTeamLogo } from '@/lib/sports-team-logo'
-import { getSportsVerticalConfig, resolveSportsVerticalFromTags } from '@/lib/sports-vertical'
+import { getSportsVerticalConfig } from '@/lib/sports-vertical'
 import { cn } from '@/lib/utils'
 import { useOrder } from '@/stores/useOrder'
 import { useSportsLivestream } from '@/stores/useSportsLivestream'
-import { useUser } from '@/stores/useUser'
-
-interface SportsGamesCenterProps {
-  cards: SportsGamesCard[]
-  sportSlug: string
-  sportTitle: string
-  pageMode?: 'games' | 'live' | 'liveAndSoon' | 'soon'
-  categoryTitleBySlug?: Record<string, string>
-  initialWeek?: number | null
-  vertical?: SportsVertical
-  showHeading?: boolean
-}
-
-type DetailsTab = 'orderBook' | 'graph' | 'about'
-export type SportsGamesMarketType = SportsGamesButton['marketType']
-export type SportsGameGraphVariant = 'default' | 'sportsCardLegend' | 'sportsEventHero'
-type LinePickerMarketType = Extract<SportsGamesMarketType, 'spread' | 'total'>
-
-const MARKET_COLUMNS: Array<{ key: SportsGamesMarketType, label: string }> = [
-  { key: 'moneyline', label: 'Moneyline' },
-  { key: 'spread', label: 'Spread' },
-  { key: 'total', label: 'Total' },
-]
-const COLLAPSED_MARKET_COLUMNS: Array<{ key: SportsGamesMarketType, label: string }> = [
-  { key: 'moneyline', label: 'Moneyline' },
-  { key: 'binary', label: 'Market' },
-  { key: 'btts', label: 'Both Teams to Score' },
-  { key: 'spread', label: 'Spread' },
-  { key: 'total', label: 'Total' },
-]
-const MARKET_COLUMN_BY_KEY = new Map(
-  [...MARKET_COLUMNS, ...COLLAPSED_MARKET_COLUMNS].map(column => [column.key, column]),
-)
-const headerIconButtonClass = `
-  flex size-10 items-center justify-center rounded-sm border border-transparent bg-transparent text-foreground
-  transition-colors
-  hover:bg-muted/80 focus-visible:ring-1 focus-visible:ring-ring md:h-9 md:w-9
-`
-const SPORTS_EVENT_ODDS_FORMAT_STORAGE_KEY = 'sports:event:odds-format'
-const SPORTS_GAMES_SHOW_SPREADS_TOTALS_STORAGE_KEY = 'sports:games:show-spreads-totals'
-const SPORTS_LIVE_FALLBACK_WINDOW_MS = 2 * 60 * 60 * 1000
-const HERO_LEGEND_LABEL_GAP_PX = 8
-const HERO_LEGEND_RIGHT_INSET_PX = 4
-const HERO_LEGEND_MIN_WIDTH_PX = 72
-const HERO_LEGEND_NAME_PADDING_PX = 10
-const HERO_LEGEND_NAME_LINE_HEIGHT_PX = 18
-const HERO_LEGEND_VALUE_LINE_HEIGHT_PX = 32
-const HERO_LEGEND_MIN_HEIGHT_PX = 56
-const HERO_LEGEND_VERTICAL_GAP_PX = 10
-const TRADE_FLOW_MAX_ITEMS = 6
-const TRADE_FLOW_TTL_MS = 8000
-const TRADE_FLOW_CLEANUP_INTERVAL_MS = 500
-const tradeFlowTextStrokeStyle = {
-  textShadow: `
-    1px 0 0 var(--background),
-    -1px 0 0 var(--background),
-    0 1px 0 var(--background),
-    0 -1px 0 var(--background),
-    1px 1px 0 var(--background),
-    -1px -1px 0 var(--background),
-    1px -1px 0 var(--background),
-    -1px 1px 0 var(--background)
-  `,
-} as const
-const GENERIC_SPORTS_CATEGORY_LABELS = new Set([
-  'esports',
-  'sports',
-  'games',
-  'live',
-  'in play',
-  'inplay',
-  'today',
-  'tomorrow',
-])
-
-function resolveInitialSportsEventOddsFormat(): OddsFormat {
-  if (typeof window === 'undefined') {
-    return 'price'
-  }
-
-  const storedOddsFormat = window.localStorage.getItem(SPORTS_EVENT_ODDS_FORMAT_STORAGE_KEY)
-  const matchedOption = ODDS_FORMAT_OPTIONS.find(option => option.value === storedOddsFormat)
-  return matchedOption?.value ?? 'price'
-}
-
-function resolveInitialShowSpreadsAndTotals() {
-  if (typeof window === 'undefined') {
-    return false
-  }
-
-  return window.localStorage.getItem(SPORTS_GAMES_SHOW_SPREADS_TOTALS_STORAGE_KEY) === '1'
-}
-
-const PredictionChart = dynamic<PredictionChartProps>(
-  () => import('@/components/PredictionChart'),
-  { ssr: false },
-)
-
-export interface SportsLinePickerOption {
-  conditionId: string
-  label: string
-  lineValue: number
-  firstIndex: number
-  buttons: SportsGamesButton[]
-}
-
-interface SportsGraphSeriesTarget {
-  key: string
-  tokenId: string | null
-  market: Market
-  outcomeIndex: number
-  name: string
-  color: string
-}
-
-interface SportsTradeFlowLabelItem {
-  id: string
-  label: string
-  color: string
-  createdAt: number
-}
-
-interface SportsTradeSelection {
-  cardId: string | null
-  buttonKey: string | null
-}
-
-interface SportsActiveTradeContext {
-  card: SportsGamesCard
-  button: SportsGamesButton
-  market: Market
-  outcome: Outcome
-}
-
-interface SportsPositionTag {
-  key: string
-  conditionId: string
-  outcomeIndex: typeof OUTCOME_INDEX.YES | typeof OUTCOME_INDEX.NO
-  marketTypeLabel: 'Moneyline' | 'Spread' | 'Total' | 'Both Teams to Score' | 'Market'
-  marketLabel: string
-  outcomeLabel: string
-  summaryLabel: string
-  shares: number
-  avgPriceCents: number | null
-  totalCost: number | null
-  currentValue: number
-  realizedPnl: number
-  market: Market
-  outcome: Outcome
-  button: SportsGamesButton | null
-  latestActivityAtMs: number
-}
-
-interface SportsCashOutModalPayload {
-  outcomeLabel: string
-  outcomeShortLabel: string
-  outcomeIconUrl: string | null | undefined
-  shares: number
-  filledShares: number | null
-  avgPriceCents: number | null
-  receiveAmount: number | null
-  sellBids: NormalizedBookLevel[]
-}
-
-function toFiniteNumber(value: unknown) {
-  if (value == null) {
-    return null
-  }
-
-  const numeric = Number(value)
-  return Number.isFinite(numeric) ? numeric : null
-}
-
-function resolvePositionShares(position: UserPosition) {
-  const quantity = toFiniteNumber(position.size)
-    ?? (typeof position.total_shares === 'number' ? position.total_shares : 0)
-  return Number.isFinite(quantity) ? quantity : 0
-}
-
-function normalizePositionPrice(value: unknown) {
-  const numeric = toFiniteNumber(value)
-  if (numeric == null || numeric <= 0) {
-    return numeric
-  }
-
-  let normalized = numeric
-  while (normalized > 1) {
-    normalized /= 100
-  }
-
-  return normalized
-}
-
-function resolvePositionCostValue(position: UserPosition, shares: number, avgPrice: number | null) {
-  const derivedCost = shares > 0 && typeof avgPrice === 'number' && avgPrice > 0 ? avgPrice * shares : null
-  if (derivedCost != null) {
-    return derivedCost
-  }
-
-  const baseCostValue = toFiniteNumber(position.totalBought)
-    ?? toFiniteNumber(position.initialValue)
-    ?? (typeof position.total_position_cost === 'number'
-      ? Number(fromMicro(String(position.total_position_cost), 2))
-      : null)
-
-  return baseCostValue
-}
-
-function resolvePositionCurrentValue(
-  position: UserPosition,
-  shares: number,
-  avgPrice: number | null,
-  marketPrice: number | null,
-) {
-  if (shares > 0) {
-    const livePrice = marketPrice ?? normalizePositionPrice(position.curPrice)
-    if (livePrice && livePrice > 0) {
-      return livePrice * shares
-    }
-  }
-
-  let value = toFiniteNumber(position.currentValue)
-    ?? Number(fromMicro(String(position.total_position_value ?? 0), 2))
-
-  if (!(value > 0) && shares > 0) {
-    if (typeof avgPrice === 'number' && avgPrice > 0) {
-      value = avgPrice * shares
-    }
-  }
-
-  return Number.isFinite(value) ? value : 0
-}
-
-function normalizePositionPnlValue(value: number | null, baseCostValue: number | null) {
-  if (!Number.isFinite(value)) {
-    return 0
-  }
-  if (!baseCostValue || baseCostValue <= 0) {
-    return value ?? 0
-  }
-  if (Math.abs(value ?? 0) <= baseCostValue * 10) {
-    return value ?? 0
-  }
-  const scaled = (value ?? 0) / MICRO_UNIT
-  if (Math.abs(scaled) <= baseCostValue * 10) {
-    return scaled
-  }
-  return 0
-}
-
-function buildTradeFlowLabel(price: number, size: number) {
-  const notional = price * size
-  if (!Number.isFinite(notional) || notional <= 0) {
-    return null
-  }
-  return formatSharePriceLabel(notional / 100, { fallback: '0¢', currencyDigits: 0 })
-}
-
-function pruneTradeFlowItems(items: SportsTradeFlowLabelItem[], now: number) {
-  return items.filter(item => now - item.createdAt <= TRADE_FLOW_TTL_MS)
-}
-
-function trimTradeFlowItems(items: SportsTradeFlowLabelItem[]) {
-  return items.slice(-TRADE_FLOW_MAX_ITEMS)
-}
-
-function resolveMarketTypeLabel(
-  button: SportsGamesButton | null,
-  market: Market,
-): 'Moneyline' | 'Spread' | 'Total' | 'Both Teams to Score' | 'Market' {
-  if (button?.marketType === 'moneyline') {
-    return 'Moneyline'
-  }
-  if (button?.marketType === 'spread') {
-    return 'Spread'
-  }
-  if (button?.marketType === 'total') {
-    return 'Total'
-  }
-  if (button?.marketType === 'btts') {
-    return 'Both Teams to Score'
-  }
-  if (button?.marketType === 'binary') {
-    return 'Market'
-  }
-
-  const normalizedType = normalizeComparableText(market.sports_market_type)
-  if (normalizedType.includes('both teams to score') || normalizedType.includes('btts')) {
-    return 'Both Teams to Score'
-  }
-  if (normalizedType.includes('spread') || normalizedType.includes('handicap')) {
-    return 'Spread'
-  }
-  if (normalizedType.includes('total') || normalizedType.includes('over under')) {
-    return 'Total'
-  }
-
-  return 'Market'
-}
-
-function formatCompactCentsLabel(cents: number | null) {
-  if (cents == null || !Number.isFinite(cents)) {
-    return '—'
-  }
-
-  const rounded = Math.round(cents * 10) / 10
-  return Number.isInteger(rounded)
-    ? `${rounded}c`
-    : `${rounded.toFixed(1)}c`
-}
-
-function normalizeHexColor(value: string | null | undefined) {
-  if (!value) {
-    return null
-  }
-
-  const trimmed = value.trim()
-  const withHash = trimmed.startsWith('#') ? trimmed : `#${trimmed}`
-  return /^#(?:[0-9a-f]{3}|[0-9a-f]{6})$/i.test(withHash) ? withHash : null
-}
-
-function resolveMoneylineButtonGridClass(buttonCount: number) {
-  if (buttonCount <= 1) {
-    return 'grid-cols-1'
-  }
-
-  if (buttonCount === 2) {
-    return 'grid-cols-2'
-  }
-
-  return 'grid-cols-3'
-}
-
-export function resolveButtonStyle(
-  color: string | null,
-  tone?: SportsGamesButton['tone'],
-): CSSProperties | undefined {
-  const normalized = normalizeHexColor(color)
-  if (!normalized) {
-    if (!isSportsTeamTone(tone)) {
-      return undefined
-    }
-
-    return resolveSportsTeamFallbackButtonStyle(tone)
-  }
-
-  return {
-    backgroundColor: normalized,
-    color: '#fff',
-  }
-}
-
-export function resolveButtonDepthStyle(
-  color: string | null,
-  tone?: SportsGamesButton['tone'],
-): CSSProperties | undefined {
-  const normalized = normalizeHexColor(color)
-  if (!normalized) {
-    if (!isSportsTeamTone(tone)) {
-      return undefined
-    }
-
-    return resolveSportsTeamFallbackDepthStyle(tone)
-  }
-
-  const hex = normalized.replace('#', '')
-  const expandedHex = hex.length === 3
-    ? hex.split('').map(char => `${char}${char}`).join('')
-    : hex
-
-  const red = Number.parseInt(expandedHex.slice(0, 2), 16)
-  const green = Number.parseInt(expandedHex.slice(2, 4), 16)
-  const blue = Number.parseInt(expandedHex.slice(4, 6), 16)
-
-  if ([red, green, blue].some(value => Number.isNaN(value))) {
-    return undefined
-  }
-
-  return {
-    backgroundColor: `rgb(${red} ${green} ${blue} / 0.8)`,
-  }
-}
-
-export function resolveButtonOverlayStyle(
-  color: string | null,
-  tone?: SportsGamesButton['tone'],
-): CSSProperties | undefined {
-  const normalized = normalizeHexColor(color)
-  if (normalized || !isSportsTeamTone(tone)) {
-    return undefined
-  }
-
-  return resolveSportsTeamFallbackOverlayStyle(tone)
-}
-
-function normalizeOutcomePriceCents(outcome: Outcome | null | undefined, market: Market) {
-  const outcomeIndex = outcome?.outcome_index === OUTCOME_INDEX.NO
-    ? OUTCOME_INDEX.NO
-    : OUTCOME_INDEX.YES
-
-  return resolveOutcomePriceCents(market, outcomeIndex) ?? 50
-}
-
-export function groupButtonsByMarketType(buttons: SportsGamesButton[]) {
-  const grouped: Record<SportsGamesMarketType, SportsGamesButton[]> = {
-    moneyline: [],
-    spread: [],
-    total: [],
-    btts: [],
-    binary: [],
-  }
-
-  for (const button of buttons) {
-    grouped[button.marketType].push(button)
-  }
-
-  return grouped
-}
-
-function toDateGroupKey(date: Date) {
-  const year = date.getFullYear()
-  const month = String(date.getMonth() + 1).padStart(2, '0')
-  const day = String(date.getDate()).padStart(2, '0')
-  return `${year}-${month}-${day}`
-}
-
-export function resolveDefaultConditionId(card: SportsGamesCard) {
-  return card.defaultConditionId
-    ?? card.buttons[0]?.key
-    ?? card.detailMarkets[0]?.condition_id
-    ?? null
-}
-
-export function resolveSelectedButton(card: SportsGamesCard, selectedButtonKey: string | null) {
-  if (selectedButtonKey) {
-    const selected = card.buttons.find(button => button.key === selectedButtonKey)
-    if (selected) {
-      return selected
-    }
-  }
-
-  return card.buttons[0] ?? null
-}
-
-export function resolveSelectedMarket(card: SportsGamesCard, selectedButtonKey: string | null) {
-  const selectedButton = resolveSelectedButton(card, selectedButtonKey)
-  if (selectedButton) {
-    const selectedMarket = card.detailMarkets.find(market => market.condition_id === selectedButton.conditionId)
-    if (selectedMarket) {
-      return selectedMarket
-    }
-  }
-
-  return card.detailMarkets[0] ?? null
-}
-
-export function resolveOrderPanelOutcomeLabelOverrides(
-  card: SportsGamesCard,
-  market: Market | null | undefined,
-): Partial<Record<number, string>> {
-  if (!market) {
-    return {}
-  }
-
-  const hasDrawMoneylineOption = card.buttons.some(button =>
-    button.marketType === 'moneyline' && button.tone === 'draw',
-  )
-  const isMoneylineCondition = card.buttons.some(button =>
-    button.conditionId === market.condition_id && button.marketType === 'moneyline',
-  )
-  if (hasDrawMoneylineOption && isMoneylineCondition) {
-    return {}
-  }
-
-  const labels: Partial<Record<number, string>> = {}
-  card.buttons.forEach((button) => {
-    if (button.conditionId !== market.condition_id || button.tone === 'draw') {
-      return
-    }
-
-    const label = button.label?.trim()
-    if (!label || labels[button.outcomeIndex]) {
-      return
-    }
-
-    labels[button.outcomeIndex] = label
-  })
-
-  return labels
-}
-
-export function resolveOrderPanelOutcomeAccentOverrides(
-  card: SportsGamesCard,
-  market: Market | null | undefined,
-): Partial<Record<number, EventOrderPanelOutcomeSelectedAccent>> {
-  if (!market) {
-    return {}
-  }
-
-  const accents: Partial<Record<number, EventOrderPanelOutcomeSelectedAccent>> = {}
-  card.buttons.forEach((button) => {
-    if (
-      button.conditionId !== market.condition_id
-      || (button.outcomeIndex !== OUTCOME_INDEX.YES && button.outcomeIndex !== OUTCOME_INDEX.NO)
-      || (button.tone !== 'team1' && button.tone !== 'team2')
-      || accents[button.outcomeIndex]
-    ) {
-      return
-    }
-
-    accents[button.outcomeIndex] = {
-      buttonStyle: resolveButtonStyle(button.color, button.tone),
-      depthStyle: resolveButtonDepthStyle(button.color, button.tone),
-      overlayStyle: resolveButtonOverlayStyle(button.color, button.tone),
-    }
-  })
-
-  return accents
-}
-
-function resolveActiveMarketType(card: SportsGamesCard, selectedButtonKey: string | null): SportsGamesMarketType {
-  if (selectedButtonKey) {
-    const selectedButton = card.buttons.find(button => button.key === selectedButtonKey)
-    if (selectedButton) {
-      return selectedButton.marketType
-    }
-  }
-
-  return card.buttons[0]?.marketType ?? 'moneyline'
-}
-
-export function resolveSelectedOutcome(market: Market | null, selectedButton: SportsGamesButton | null): Outcome | null {
-  if (!market) {
-    return null
-  }
-
-  if (selectedButton) {
-    const selectedOutcome = market.outcomes.find(outcome => outcome.outcome_index === selectedButton.outcomeIndex)
-    if (selectedOutcome) {
-      return selectedOutcome
-    }
-  }
-
-  return market.outcomes.find(outcome => outcome.outcome_index === OUTCOME_INDEX.YES)
-    ?? market.outcomes[0]
-    ?? null
-}
-
-export function resolveStableSpreadPrimaryOutcomeIndex(card: SportsGamesCard, conditionId: string) {
-  const spreadButtonsForCondition = card.buttons
-    .filter(button => button.marketType === 'spread' && button.conditionId === conditionId)
-    .map(button => button.outcomeIndex)
-    .filter((index): index is number => index === OUTCOME_INDEX.YES || index === OUTCOME_INDEX.NO)
-  const uniqueButtonIndices = Array.from(new Set(spreadButtonsForCondition)).sort((a, b) => a - b)
-
-  // Spread in sports is rendered with inverted side order in the order panel
-  // compared to outcome index ordering, so pick the opposite side as primary.
-  if (uniqueButtonIndices.length >= 2) {
-    return uniqueButtonIndices[1]
-  }
-  if (uniqueButtonIndices.length === 1) {
-    return uniqueButtonIndices[0]
-  }
-
-  const market = card.detailMarkets.find(item => item.condition_id === conditionId)
-  if (!market) {
-    return null
-  }
-
-  const marketIndices = Array.from(market.outcomes, outcome => outcome.outcome_index)
-    .filter((index): index is number => index === OUTCOME_INDEX.YES || index === OUTCOME_INDEX.NO)
-  const uniqueMarketIndices = Array.from(new Set(marketIndices)).sort((a, b) => a - b)
-
-  if (uniqueMarketIndices.length >= 2) {
-    return uniqueMarketIndices[1]
-  }
-  if (uniqueMarketIndices.length === 1) {
-    return uniqueMarketIndices[0]
-  }
-
-  return null
-}
-
-function extractLineValue(value: string) {
-  const match = value.match(/([+-]?\d+(?:\.\d+)?)/)
-  return match?.[1] ?? null
-}
-
-function formatLineValue(value: number) {
-  const rounded = Math.round(value * 10) / 10
-  return Number.isInteger(rounded) ? `${rounded}` : `${rounded}`
-}
-
-function toLineNumber(value: string | null) {
-  if (!value) {
-    return null
-  }
-
-  const parsed = Number(value)
-  return Number.isFinite(parsed) ? Math.abs(parsed) : null
-}
-
-function resolveMarketLineValue(market: Market | null, marketType: LinePickerMarketType) {
-  if (!market) {
-    return null
-  }
-
-  const marketText = [
-    market.sports_group_item_title,
-    market.short_title,
-    market.title,
-    ...market.outcomes.map(outcome => outcome.outcome_text),
-  ]
-    .filter((value): value is string => Boolean(value?.trim()))
-    .join(' ')
-
-  const rawLine = extractLineValue(marketText)
-  if (!rawLine) {
-    return null
-  }
-
-  const lineValue = toLineNumber(rawLine)
-  if (lineValue === null) {
-    return null
-  }
-
-  return marketType === 'spread'
-    ? lineValue
-    : lineValue
-}
-
-export function buildLinePickerOptions(card: SportsGamesCard, marketType: LinePickerMarketType): SportsLinePickerOption[] {
-  const sourceButtons = card.buttons.filter(button => button.marketType === marketType)
-  if (sourceButtons.length === 0) {
-    return []
-  }
-
-  const marketByConditionId = new Map(card.detailMarkets.map(market => [market.condition_id, market] as const))
-  const byCondition = new Map<string, SportsLinePickerOption>()
-
-  sourceButtons.forEach((button, index) => {
-    const existing = byCondition.get(button.conditionId)
-    if (existing) {
-      existing.buttons.push(button)
-      return
-    }
-
-    const market = marketByConditionId.get(button.conditionId) ?? null
-    const fromMarket = resolveMarketLineValue(market, marketType)
-    const fromButton = toLineNumber(extractLineValue(button.label))
-    const lineValue = fromMarket ?? fromButton
-    if (lineValue === null) {
-      return
-    }
-
-    byCondition.set(button.conditionId, {
-      conditionId: button.conditionId,
-      label: formatLineValue(lineValue),
-      lineValue,
-      firstIndex: index,
-      buttons: [button],
-    })
-  })
-
-  return Array.from(byCondition.values())
-    .sort((a, b) => {
-      if (a.lineValue !== b.lineValue) {
-        return a.lineValue - b.lineValue
-      }
-      return a.firstIndex - b.firstIndex
-    })
-}
-
-export function resolvePreferredLinePickerButton(
-  buttons: SportsGamesButton[],
-  selectedButton: SportsGamesButton | null | undefined,
-) {
-  if (buttons.length === 0) {
-    return null
-  }
-
-  if (!selectedButton) {
-    return buttons[0] ?? null
-  }
-
-  if (selectedButton.tone !== 'neutral') {
-    const toneMatch = buttons.find(button => button.tone === selectedButton.tone)
-    if (toneMatch) {
-      return toneMatch
-    }
-  }
-
-  const outcomeIndexMatch = buttons.find(button => button.outcomeIndex === selectedButton.outcomeIndex)
-  if (outcomeIndexMatch) {
-    return outcomeIndexMatch
-  }
-
-  return buttons[0] ?? null
-}
-
-function resolveGraphSeriesName(card: SportsGamesCard, button: SportsGamesButton | undefined, market: Market) {
-  if (!button) {
-    return market.sports_group_item_title?.trim()
-      || market.short_title?.trim()
-      || market.title
-  }
-
-  if (button.tone === 'team1') {
-    return card.teams[0]?.name ?? button.label
-  }
-  if (button.tone === 'team2') {
-    return card.teams[1]?.name ?? button.label
-  }
-  if (button.tone === 'draw') {
-    return 'Draw'
-  }
-
-  return button.label
-}
-
-function resolveGraphSeriesColor(
-  card: SportsGamesCard,
-  button: SportsGamesButton | undefined,
-  fallbackColor: string,
-) {
-  const relatedColor = normalizeHexColor(button?.color)
-  if (relatedColor) {
-    return relatedColor
-  }
-
-  const relatedTeamColor = normalizeHexColor(
-    button ? resolveTeamByTone(card, button.tone)?.color : null,
-  )
-  if (relatedTeamColor) {
-    return relatedTeamColor
-  }
-  if (button && isSportsTeamTone(button.tone)) {
-    return resolveSportsTeamFallbackColor(button.tone)
-  }
-
-  if (button?.tone === 'over') {
-    return 'var(--yes)'
-  }
-  if (button?.tone === 'under') {
-    return 'var(--no)'
-  }
-  if (button?.tone === 'draw') {
-    return 'var(--secondary-foreground)'
-  }
-
-  return fallbackColor
-}
-
-function buildCompositeMoneylineGraphTargets(card: SportsGamesCard) {
-  const moneylineButtons = card.buttons.filter(button => button.marketType === 'moneyline')
-  if (moneylineButtons.length < 2) {
-    return [] as SportsGraphSeriesTarget[]
-  }
-
-  const moneylineConditionIds = Array.from(new Set(moneylineButtons.map(button => button.conditionId)))
-  if (moneylineConditionIds.length !== 1) {
-    return [] as SportsGraphSeriesTarget[]
-  }
-
-  const market = card.detailMarkets.find(
-    detailMarket => detailMarket.condition_id === moneylineConditionIds[0],
-  ) ?? null
-  if (!market) {
-    return [] as SportsGraphSeriesTarget[]
-  }
-
-  const orderedButtons = [...moneylineButtons].sort((left, right) => {
-    const toneOrder: Record<SportsGamesButton['tone'], number> = {
-      team1: 0,
-      draw: 1,
-      team2: 2,
-      over: 3,
-      under: 4,
-      neutral: 5,
-    }
-
-    return (toneOrder[left.tone] ?? 99) - (toneOrder[right.tone] ?? 99)
-  })
-
-  const fallbackColors = ['var(--yes)', 'var(--secondary-foreground)', 'var(--no)']
-
-  return orderedButtons.reduce<SportsGraphSeriesTarget[]>((targets, button, index) => {
-    const outcome = market.outcomes.find(
-      candidate => candidate.outcome_index === button.outcomeIndex,
-    ) ?? null
-
-    if (!outcome?.token_id) {
-      return targets
-    }
-
-    targets.push({
-      key: `${market.condition_id}:${button.outcomeIndex}`,
-      tokenId: outcome.token_id,
-      market,
-      outcomeIndex: button.outcomeIndex,
-      name: resolveGraphSeriesName(card, button, market),
-      color: resolveGraphSeriesColor(card, button, fallbackColors[index % fallbackColors.length]!),
-    })
-
-    return targets
-  }, [])
-}
-
-function resolveTotalButtonLabel(button: SportsGamesButton, selectedOutcome: Outcome | null) {
-  const line = extractLineValue(button.label)
-  const outcomeText = selectedOutcome?.outcome_text?.trim() ?? ''
-
-  let sideLabel: 'OVER' | 'UNDER'
-  if (/^under$/i.test(outcomeText) || button.tone === 'under') {
-    sideLabel = 'UNDER'
-  }
-  else if (/^over$/i.test(outcomeText) || button.tone === 'over') {
-    sideLabel = 'OVER'
-  }
-  else {
-    sideLabel = button.label.trim().toUpperCase().startsWith('U') ? 'UNDER' : 'OVER'
-  }
-
-  return line ? `${sideLabel} ${line}` : sideLabel
-}
-
-function resolveSelectedTradeLabel(
-  card: SportsGamesCard,
-  button: SportsGamesButton | null,
-  selectedOutcome: Outcome | null,
-) {
-  if (!button) {
-    return selectedOutcome?.outcome_text?.trim().toUpperCase() || 'YES'
-  }
-
-  if (button.marketType === 'total') {
-    return resolveTotalButtonLabel(button, selectedOutcome)
-  }
-
-  if (button.marketType === 'moneyline' && (button.tone === 'team1' || button.tone === 'team2')) {
-    const teamName = resolveTeamByTone(card, button.tone)?.name?.trim()
-    if (teamName) {
-      return teamName
-    }
-  }
-
-  return button.label.trim().toUpperCase()
-}
-
-function resolveMarketDescriptor(market: Market | null) {
-  if (!market) {
-    return null
-  }
-
-  const descriptor = market.sports_group_item_title?.trim()
-    || market.short_title?.trim()
-    || market.title?.trim()
-    || ''
-  return descriptor || null
-}
-
-function normalizeComparableText(value: string | null | undefined) {
-  return value
-    ?.normalize('NFKD')
-    .replace(/[\u0300-\u036F]/g, '')
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, ' ')
-    .trim()
-    ?? ''
-}
-
-function toFiniteTimestamp(value: string | null | undefined) {
-  if (!value) {
-    return Number.NaN
-  }
-
-  const timestamp = Date.parse(value)
-  return Number.isFinite(timestamp) ? timestamp : Number.NaN
-}
-
-function resolveCardStartTimestamp(card: SportsGamesCard) {
-  return toFiniteTimestamp(
-    card.startTime
-    ?? card.event.sports_start_time
-    ?? card.event.start_date,
-  )
-}
-
-function resolveCardEndTimestamp(card: SportsGamesCard) {
-  const explicitEnd = toFiniteTimestamp(card.event.end_date)
-  if (Number.isFinite(explicitEnd)) {
-    return explicitEnd
-  }
-
-  const marketEndTimes = card.detailMarkets
-    .map(market => toFiniteTimestamp(market.end_time ?? null))
-    .filter(timestamp => Number.isFinite(timestamp))
-
-  if (marketEndTimes.length > 0) {
-    return Math.max(...marketEndTimes)
-  }
-
-  return Number.NaN
-}
-
-function resolveCardLiveFallbackEndTimestamp(card: SportsGamesCard) {
-  const startMs = resolveCardStartTimestamp(card)
-  if (!Number.isFinite(startMs)) {
-    return Number.NaN
-  }
-
-  const endMs = resolveCardEndTimestamp(card)
-  const referenceEndMs = Number.isFinite(endMs) && endMs > startMs
-    ? endMs
-    : startMs
-
-  return referenceEndMs + SPORTS_LIVE_FALLBACK_WINDOW_MS
-}
-
-function parseSportsScore(value: string | null | undefined) {
-  const trimmed = value?.trim()
-  if (!trimmed) {
-    return null
-  }
-
-  const match = trimmed.match(/(\d+)\D+(\d+)/)
-  if (!match) {
-    return null
-  }
-
-  const team1 = Number.parseInt(match[1] ?? '', 10)
-  const team2 = Number.parseInt(match[2] ?? '', 10)
-  if (!Number.isFinite(team1) || !Number.isFinite(team2)) {
-    return null
-  }
-
-  return { team1, team2 }
-}
-
-function isCardLiveNow(card: SportsGamesCard, nowMs: number) {
-  if (card.event.status !== 'active' || card.event.sports_ended === true) {
-    return false
-  }
-
-  const startMs = resolveCardStartTimestamp(card)
-  const endMs = resolveCardEndTimestamp(card)
-  const isInTimeWindow = Number.isFinite(startMs) && Number.isFinite(endMs)
-    ? startMs <= nowMs && nowMs <= endMs
-    : false
-  const liveFallbackEndMs = resolveCardLiveFallbackEndTimestamp(card)
-  const isWithinFallbackWindow = Number.isFinite(startMs) && Number.isFinite(liveFallbackEndMs)
-    ? startMs <= nowMs && nowMs <= liveFallbackEndMs
-    : false
-
-  if (card.event.sports_live === true) {
-    return true
-  }
-
-  return isInTimeWindow || isWithinFallbackWindow
-}
-
-function isCardFuture(card: SportsGamesCard, nowMs: number) {
-  if (card.event.status !== 'active') {
-    return false
-  }
-
-  const startMs = resolveCardStartTimestamp(card)
-  return Number.isFinite(startMs) && startMs > nowMs
-}
-
-function formatCategoryFromSlug(value: string) {
-  return value
-    .split('-')
-    .filter(Boolean)
-    .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(' ')
-}
-
-function resolveCategoryFromSportSlugs(card: SportsGamesCard) {
-  const categorySlug = card.event.sports_sport_slug?.trim() || card.event.sports_series_slug?.trim()
-  if (!categorySlug) {
-    return null
-  }
-
-  return formatCategoryFromSlug(categorySlug)
-}
-
-function resolveCategoryFromEventSlug(card: SportsGamesCard) {
-  const eventSlug = card.event.sports_event_slug?.trim() || card.event.series_slug?.trim()
-  if (!eventSlug) {
-    return null
-  }
-
-  const cleaned = eventSlug
-    .replace(/-games?$/i, '')
-    .replace(/-live$/i, '')
-    .replace(/-props$/i, '')
-  if (!cleaned) {
-    return null
-  }
-
-  const sportSlug = card.event.sports_sport_slug?.trim().toLowerCase()
-  const tokens = cleaned.split('-').filter(Boolean)
-  const normalizedTokens = sportSlug
-    ? tokens.filter(token => token.toLowerCase() !== sportSlug)
-    : tokens
-  const candidate = normalizedTokens.join('-')
-
-  return candidate ? formatCategoryFromSlug(candidate) : null
-}
-
-function isGenericSportsCategoryLabel(label: string, sportSlug: string | null | undefined) {
-  const normalized = normalizeComparableText(label)
-  if (!normalized) {
-    return true
-  }
-
-  if (GENERIC_SPORTS_CATEGORY_LABELS.has(normalized)) {
-    return true
-  }
-
-  if (!sportSlug) {
-    return false
-  }
-
-  const normalizedSportSlug = normalizeComparableText(sportSlug.replace(/-/g, ' '))
-  return normalized === normalizedSportSlug
-}
-
-function resolveCardCategoryLabel(
-  card: SportsGamesCard,
-  categoryTitleBySlug: Record<string, string> = {},
-) {
-  const categorySlugCandidates = [
-    card.event.sports_series_slug,
-    card.event.series_slug,
-    card.event.sports_sport_slug,
-  ]
-    .map(value => value?.trim().toLowerCase() ?? '')
-    .filter(Boolean)
-
-  for (const slug of categorySlugCandidates) {
-    const mappedCategoryTitle = categoryTitleBySlug[slug]
-    if (
-      mappedCategoryTitle
-      && !isGenericSportsCategoryLabel(mappedCategoryTitle, slug)
-    ) {
-      return mappedCategoryTitle
-    }
-  }
-
-  const genericSlug = categorySlugCandidates[0] ?? card.event.sports_sport_slug
-  const candidateTags = card.event.sports_tags
-    ?.map(tag => tag?.trim() ?? '')
-    .filter(Boolean)
-    .filter(tag => !isGenericSportsCategoryLabel(tag, genericSlug))
-    ?? []
-
-  if (candidateTags.length > 0) {
-    return [...candidateTags].sort((a, b) => b.length - a.length)[0]!
-  }
-
-  return resolveCategoryFromSportSlugs(card)
-    ?? resolveCategoryFromEventSlug(card)
-    ?? 'Other'
-}
-
-function resolveSwitchTooltip(market: Market | null, nextOutcome: Outcome | null) {
-  if (!nextOutcome) {
-    return null
-  }
-
-  const nextOutcomeLabel = nextOutcome.outcome_text?.trim() || null
-  if (!nextOutcomeLabel) {
-    return null
-  }
-
-  const marketDescriptor = resolveMarketDescriptor(market)
-  if (!marketDescriptor) {
-    return `Switch to ${nextOutcomeLabel}`
-  }
-
-  const normalizedOutcome = normalizeComparableText(nextOutcomeLabel)
-  const normalizedDescriptor = normalizeComparableText(marketDescriptor)
-  if (!normalizedDescriptor || normalizedDescriptor === normalizedOutcome) {
-    return `Switch to ${nextOutcomeLabel}`
-  }
-
-  return `Switch to ${nextOutcomeLabel} - ${marketDescriptor}`
-}
-
-function useSportsGameGraphChartSettings() {
-  const [chartSettings, setChartSettings] = useState(function resolveInitialGraphChartSettings() {
-    const stored = loadStoredChartSettings()
-    return { ...stored, bothOutcomes: false }
-  })
-
-  useEffect(function persistGraphChartSettings() {
-    storeChartSettings({ ...chartSettings, bothOutcomes: false })
-    return function noopGraphChartSettingsCleanup() {}
-  }, [chartSettings])
-
-  return [chartSettings, setChartSettings] as const
-}
-
-function useSportsGameGraphChartDimensions({
-  windowWidth,
-  variant,
-}: {
-  windowWidth: number | undefined
-  variant: SportsGameGraphVariant
-}) {
-  const isSportsEventHeroVariant = variant === 'sportsEventHero'
-  const usesPositionedSeriesLegend = variant === 'sportsEventHero' || variant === 'sportsCardLegend'
-  const chartHeight = isSportsEventHeroVariant ? 332 : 300
-  const chartMargin = usesPositionedSeriesLegend
-    ? { top: 12, right: 46, bottom: 40, left: 0 }
-    : { top: 12, right: 30, bottom: 40, left: 0 }
-
-  const chartWidth = useMemo(() => {
-    const viewportWidth = windowWidth ?? 1200
-
-    if (viewportWidth < 768) {
-      return Math.max(260, viewportWidth - 112)
-    }
-
-    return Math.min(860, viewportWidth - 520)
-  }, [windowWidth])
-
-  return {
-    isSportsEventHeroVariant,
-    usesPositionedSeriesLegend,
-    canRenderPositionedSeriesLegend: usesPositionedSeriesLegend,
-    chartHeight,
-    chartMargin,
-    chartWidth,
-  }
-}
-
-function useSportsGameGraphSeries({
-  card,
-  selectedConditionId,
-  isSecondaryMarketGraph,
-  isSportsEventHeroVariant,
-}: {
-  card: SportsGamesCard
-  selectedConditionId: string | null
-  isSecondaryMarketGraph: boolean
-  isSportsEventHeroVariant: boolean
-}) {
-  const graphSeriesTargets = useMemo<SportsGraphSeriesTarget[]>(
-    () => {
-      if (
-        selectedConditionId
-        && isSecondaryMarketGraph
-      ) {
-        const selectedMarket = card.detailMarkets.find(
-          market => market.condition_id === selectedConditionId,
-        )
-        if (selectedMarket) {
-          const fallbackColors = ['var(--yes)', 'var(--no)']
-          const orderedOutcomes = [...selectedMarket.outcomes]
-            .sort((a, b) => a.outcome_index - b.outcome_index)
-
-          const outcomeTargets = orderedOutcomes
-            .map((outcome, index) => {
-              const relatedButton = card.buttons.find(
-                button => button.conditionId === selectedMarket.condition_id
-                  && button.outcomeIndex === outcome.outcome_index,
-              )
-              const fallbackLabel = outcome.outcome_text?.trim() || `Option ${index + 1}`
-
-              return {
-                key: `${selectedMarket.condition_id}:${outcome.outcome_index}`,
-                tokenId: outcome.token_id ?? null,
-                market: selectedMarket,
-                outcomeIndex: outcome.outcome_index,
-                name: relatedButton?.label ?? fallbackLabel,
-                color: resolveGraphSeriesColor(card, relatedButton, fallbackColors[index % fallbackColors.length]!),
-              }
-            })
-
-          if (outcomeTargets.length > 0) {
-            return outcomeTargets
-          }
-        }
-      }
-
-      const fallbackColors = ['var(--yes)', 'var(--primary)', 'var(--no)']
-
-      const compositeMoneylineTargets = buildCompositeMoneylineGraphTargets(card)
-      if (compositeMoneylineTargets.length > 0) {
-        return compositeMoneylineTargets
-      }
-
-      const moneylineConditionIds = Array.from(new Set(
-        card.buttons
-          .filter(button => button.marketType === 'moneyline')
-          .map(button => button.conditionId),
-      ))
-
-      const moneylineMarkets = moneylineConditionIds
-        .map(conditionId => card.detailMarkets.find(market => market.condition_id === conditionId) ?? null)
-        .filter((market): market is Market => Boolean(market))
-
-      if (moneylineMarkets.length > 0) {
-        return moneylineMarkets
-          .map<SportsGraphSeriesTarget | null>((market, index) => {
-            const yesOutcome = market.outcomes.find(outcome => outcome.outcome_index === OUTCOME_INDEX.YES)
-              ?? market.outcomes[0]
-              ?? null
-            if (!yesOutcome?.token_id) {
-              return null
-            }
-
-            const relatedButton = card.buttons.find(
-              button => button.conditionId === market.condition_id
-                && button.outcomeIndex === yesOutcome.outcome_index,
-            ) ?? card.buttons.find(button => button.conditionId === market.condition_id)
-
-            return {
-              key: market.condition_id,
-              tokenId: yesOutcome.token_id,
-              market,
-              outcomeIndex: yesOutcome.outcome_index,
-              name: resolveGraphSeriesName(card, relatedButton, market),
-              color: resolveGraphSeriesColor(card, relatedButton, fallbackColors[index % fallbackColors.length]!),
-            }
-          })
-          .filter((target): target is SportsGraphSeriesTarget => target !== null)
-      }
-
-      const seenConditionIds = new Set<string>()
-      const fallbackTargets: SportsGraphSeriesTarget[] = []
-      for (const market of card.detailMarkets) {
-        if (seenConditionIds.has(market.condition_id)) {
-          continue
-        }
-        seenConditionIds.add(market.condition_id)
-        const yesOutcome = market.outcomes.find(outcome => outcome.outcome_index === OUTCOME_INDEX.YES)
-          ?? market.outcomes[0]
-          ?? null
-        if (!yesOutcome?.token_id) {
-          continue
-        }
-
-        const relatedButton = card.buttons.find(
-          button => button.conditionId === market.condition_id
-            && button.outcomeIndex === yesOutcome.outcome_index,
-        ) ?? card.buttons.find(button => button.conditionId === market.condition_id)
-
-        fallbackTargets.push({
-          key: market.condition_id,
-          tokenId: yesOutcome.token_id,
-          market,
-          outcomeIndex: yesOutcome.outcome_index,
-          name: resolveGraphSeriesName(card, relatedButton, market),
-          color: resolveGraphSeriesColor(card, relatedButton, fallbackColors[fallbackTargets.length % fallbackColors.length]!),
-        })
-      }
-
-      return fallbackTargets
-    },
-    [card, isSecondaryMarketGraph, selectedConditionId],
-  )
-
-  const tradeFlowSeriesByTokenId = useMemo(() => {
-    const map = new Map<string, { color: string }>()
-    if (!isSportsEventHeroVariant) {
-      return map
-    }
-
-    for (const series of graphSeriesTargets) {
-      if (!series.tokenId) {
-        continue
-      }
-      map.set(String(series.tokenId), {
-        color: series.color,
-      })
-    }
-
-    return map
-  }, [graphSeriesTargets, isSportsEventHeroVariant])
-
-  const marketTargets = useMemo(
-    () => graphSeriesTargets
-      .filter((target): target is SportsGraphSeriesTarget & { tokenId: string } => Boolean(target.tokenId))
-      .map(target => ({
-        conditionId: target.key,
-        tokenId: target.tokenId,
-      })),
-    [graphSeriesTargets],
-  )
-
-  const chartSeries = useMemo(() => {
-    return graphSeriesTargets.map(target => ({
-      key: target.key,
-      name: target.name,
-      color: target.color,
-    }))
-  }, [graphSeriesTargets])
-
-  return { graphSeriesTargets, tradeFlowSeriesByTokenId, marketTargets, chartSeries }
-}
-
-function useSportsGameGraphHistory({
-  card,
-  marketTargets,
-  activeTimeRange,
-  chartSeries,
-  graphSeriesTargets,
-  isSecondaryMarketGraph,
-}: {
-  card: SportsGamesCard
-  marketTargets: Array<{ conditionId: string, tokenId: string }>
-  activeTimeRange: (typeof TIME_RANGES)[number]
-  chartSeries: Array<{ key: string, name: string, color: string }>
-  graphSeriesTargets: SportsGraphSeriesTarget[]
-  isSecondaryMarketGraph: boolean
-}) {
-  const { normalizedHistory } = useEventPriceHistory({
-    eventId: card.id,
-    range: activeTimeRange,
-    targets: marketTargets,
-    eventCreatedAt: card.eventCreatedAt,
-    eventResolvedAt: card.eventResolvedAt,
-  })
-  const leadingGapStart = normalizedHistory[0]?.date ?? null
-
-  const historyChartData = useMemo<DataPoint[]>(() => {
-    return normalizedHistory
-      .map((point) => {
-        const nextPoint: DataPoint = { date: point.date }
-        let hasValue = false
-
-        for (const series of chartSeries) {
-          const value = point[series.key]
-          if (typeof value !== 'number' || !Number.isFinite(value)) {
-            continue
-          }
-
-          nextPoint[series.key] = value
-          hasValue = true
-        }
-
-        return hasValue ? nextPoint : null
-      })
-      .filter((point): point is DataPoint => point !== null)
-  }, [chartSeries, normalizedHistory])
-
-  const pairedHistoryChartData = useMemo<DataPoint[]>(() => {
-    if (!isSecondaryMarketGraph || chartSeries.length !== 2) {
-      return historyChartData
-    }
-
-    const [firstSeries, secondSeries] = chartSeries
-    return historyChartData
-      .map((point) => {
-        const firstRaw = point[firstSeries.key]
-        const secondRaw = point[secondSeries.key]
-        const firstValue = typeof firstRaw === 'number' && Number.isFinite(firstRaw) ? firstRaw : null
-        const secondValue = typeof secondRaw === 'number' && Number.isFinite(secondRaw) ? secondRaw : null
-
-        if (firstValue === null && secondValue === null) {
-          return null
-        }
-
-        const nextPoint: DataPoint = { ...point }
-        if (firstValue !== null && secondValue === null) {
-          nextPoint[secondSeries.key] = Math.max(0, Math.min(100, 100 - firstValue))
-        }
-        else if (firstValue === null && secondValue !== null) {
-          nextPoint[firstSeries.key] = Math.max(0, Math.min(100, 100 - secondValue))
-        }
-
-        return nextPoint
-      })
-      .filter((point): point is DataPoint => point !== null)
-  }, [chartSeries, historyChartData, isSecondaryMarketGraph])
-
-  const fallbackChartData = useMemo<DataPoint[]>(() => {
-    if (graphSeriesTargets.length === 0) {
-      return []
-    }
-
-    const createdMs = Date.parse(card.eventCreatedAt)
-    const resolvedMs = card.eventResolvedAt ? Date.parse(card.eventResolvedAt) : Number.NaN
-    const anchorMs = Number.isFinite(resolvedMs)
-      ? resolvedMs
-      : (Number.isFinite(createdMs) ? createdMs : Date.parse('2020-01-01T00:00:00.000Z'))
-    const endMs = anchorMs + 60_000
-    const startMs = anchorMs - (30 * 60_000)
-
-    const startPoint: DataPoint = { date: new Date(startMs) }
-    const endPoint: DataPoint = { date: new Date(endMs) }
-
-    for (const series of graphSeriesTargets) {
-      const matchingOutcome = series.market.outcomes.find(
-        outcome => outcome.outcome_index === series.outcomeIndex,
-      )
-      const cents = normalizeOutcomePriceCents(matchingOutcome, series.market)
-      startPoint[series.key] = cents
-      endPoint[series.key] = cents
-    }
-
-    return [startPoint, endPoint]
-  }, [card.eventCreatedAt, card.eventResolvedAt, graphSeriesTargets])
-
-  const chartData = pairedHistoryChartData.length > 0 ? pairedHistoryChartData : fallbackChartData
-
-  const latestSnapshot = useMemo(() => {
-    const nextValues: Record<string, number> = {}
-
-    chartSeries.forEach((seriesItem) => {
-      for (let index = chartData.length - 1; index >= 0; index -= 1) {
-        const point = chartData[index]
-        if (!point) {
-          continue
-        }
-
-        const value = point[seriesItem.key]
-        if (typeof value === 'number' && Number.isFinite(value)) {
-          nextValues[seriesItem.key] = value
-          break
-        }
-      }
-    })
-
-    return nextValues
-  }, [chartData, chartSeries])
-
-  return { chartData, latestSnapshot, leadingGapStart }
-}
-
-function useSportsGameGraphHeroLegend({
-  canRenderPositionedSeriesLegend,
-  chartSeries,
-  chartData,
-  chartWidth,
-  chartHeight,
-  chartMargin,
-  cursorSnapshot,
-  latestSnapshot,
-  usesPositionedSeriesLegend,
-}: {
-  canRenderPositionedSeriesLegend: boolean
-  chartSeries: Array<{ key: string, name: string, color: string }>
-  chartData: DataPoint[]
-  chartWidth: number
-  chartHeight: number
-  chartMargin: { top: number, right: number, bottom: number, left: number }
-  cursorSnapshot: PredictionChartCursorSnapshot | null
-  latestSnapshot: Record<string, number>
-  usesPositionedSeriesLegend: boolean
-}) {
-  const heroLegendRenderedWidth = useMemo(() => {
-    if (!canRenderPositionedSeriesLegend || chartSeries.length === 0) {
-      return HERO_LEGEND_MIN_WIDTH_PX
-    }
-
-    if (typeof document === 'undefined') {
-      return HERO_LEGEND_MIN_WIDTH_PX
-    }
-
-    const context = document.createElement('canvas').getContext('2d')
-    if (!context) {
-      return HERO_LEGEND_MIN_WIDTH_PX
-    }
-
-    context.font = '500 13px ui-sans-serif, system-ui, -apple-system, "Segoe UI", sans-serif'
-
-    const longestLabelWidth = chartSeries.reduce((maxWidth, seriesItem) => {
-      const label = seriesItem.name.trim()
-      if (!label) {
-        return maxWidth
-      }
-
-      return Math.max(maxWidth, context.measureText(label).width)
-    }, 0)
-
-    const targetWidth = Math.ceil(longestLabelWidth + HERO_LEGEND_NAME_PADDING_PX)
-    return Math.max(HERO_LEGEND_MIN_WIDTH_PX, targetWidth)
-  }, [canRenderPositionedSeriesLegend, chartSeries])
-
-  const chartXDomain = useMemo(() => {
-    if (!usesPositionedSeriesLegend || chartData.length < 2) {
-      return undefined
-    }
-
-    const firstPoint = chartData[0]
-    const lastPoint = chartData.at(-1)
-    if (!firstPoint || !lastPoint) {
-      return undefined
-    }
-
-    const firstTimestamp = firstPoint.date.getTime()
-    const lastTimestamp = lastPoint.date.getTime()
-    if (!Number.isFinite(firstTimestamp) || !Number.isFinite(lastTimestamp) || lastTimestamp <= firstTimestamp) {
-      return undefined
-    }
-
-    const dataSpanMs = Math.max(1, lastTimestamp - firstTimestamp)
-    const plotWidthPx = Math.max(1, chartWidth - chartMargin.left - chartMargin.right)
-    const reservedRightPx = Math.max(0, heroLegendRenderedWidth + HERO_LEGEND_LABEL_GAP_PX + HERO_LEGEND_RIGHT_INSET_PX)
-
-    // Keep enough fixed room on the right for legend so the plotted line ends before chart edge.
-    if (reservedRightPx >= plotWidthPx - 1) {
-      return {
-        start: firstTimestamp,
-        end: lastTimestamp,
-      }
-    }
-
-    const domainSpanMs = Math.round((dataSpanMs * plotWidthPx) / (plotWidthPx - reservedRightPx))
-    return {
-      start: firstTimestamp,
-      end: firstTimestamp + domainSpanMs,
-    }
-  }, [
-    chartData,
-    chartMargin.left,
-    chartMargin.right,
-    chartWidth,
-    heroLegendRenderedWidth,
-    usesPositionedSeriesLegend,
-  ])
-
-  const heroLegendSeriesWithValues = useMemo(
-    () => {
-      if (!canRenderPositionedSeriesLegend) {
-        return []
-      }
-
-      return chartSeries
-        .map((seriesItem) => {
-          const hoveredValue = cursorSnapshot?.values?.[seriesItem.key]
-          const value = typeof hoveredValue === 'number' && Number.isFinite(hoveredValue)
-            ? hoveredValue
-            : latestSnapshot[seriesItem.key]
-          if (typeof value !== 'number' || !Number.isFinite(value)) {
-            return null
-          }
-
-          return { ...seriesItem, value }
-        })
-        .filter((entry): entry is { key: string, name: string, color: string, value: number } => entry !== null)
-    },
-    [canRenderPositionedSeriesLegend, chartSeries, cursorSnapshot, latestSnapshot],
-  )
-
-  const heroLegendPositionedEntries = useMemo(
-    () => {
-      if (!canRenderPositionedSeriesLegend || heroLegendSeriesWithValues.length === 0 || chartData.length === 0) {
-        return [] as Array<{
-          key: string
-          name: string
-          color: string
-          value: number
-          left: number
-          top: number
-        }>
-      }
-
-      const firstPoint = chartData[0]
-      const lastPoint = chartData.at(-1)
-      if (!firstPoint || !lastPoint) {
-        return []
-      }
-
-      const firstTimestamp = firstPoint.date.getTime()
-      const lastTimestamp = lastPoint.date.getTime()
-      if (!Number.isFinite(firstTimestamp) || !Number.isFinite(lastTimestamp)) {
-        return []
-      }
-
-      const explicitStart = typeof chartXDomain?.start === 'number'
-        ? chartXDomain.start
-        : Number.NaN
-      const explicitEnd = typeof chartXDomain?.end === 'number'
-        ? chartXDomain.end
-        : Number.NaN
-      const domainStart = Number.isFinite(explicitStart) ? explicitStart : firstTimestamp
-      const domainEndCandidate = Number.isFinite(explicitEnd) ? explicitEnd : lastTimestamp
-      const domainEnd = Math.max(domainStart + 1, domainEndCandidate)
-      const hoveredTimestampRaw = cursorSnapshot?.date.getTime() ?? lastTimestamp
-      const hoveredTimestamp = Math.max(firstTimestamp, Math.min(lastTimestamp, hoveredTimestampRaw))
-
-      const yBounds = calculateYAxisBounds(chartData, chartSeries)
-      const ySpan = Math.max(1, yBounds.max - yBounds.min)
-      const xSpan = Math.max(1, domainEnd - domainStart)
-      const plotWidth = Math.max(1, chartWidth - chartMargin.left - chartMargin.right)
-      const plotHeight = Math.max(1, chartHeight - chartMargin.top - chartMargin.bottom)
-      const chartTop = chartMargin.top
-      const chartBottom = chartMargin.top + plotHeight
-      const dotX = chartMargin.left + ((hoveredTimestamp - domainStart) / xSpan) * plotWidth
-      const plotLeft = chartMargin.left
-      const plotRight = chartWidth - chartMargin.right
-      const availableFullWidth = plotRight - plotLeft - HERO_LEGEND_RIGHT_INSET_PX
-      const effectiveLabelWidth = Math.max(0, Math.min(heroLegendRenderedWidth, availableFullWidth))
-      const maxLeft = plotRight - effectiveLabelWidth - HERO_LEGEND_RIGHT_INSET_PX
-      const labelLeft = Math.max(plotLeft, Math.min(maxLeft, dotX + HERO_LEGEND_LABEL_GAP_PX))
-      const availableLabelWidth = Math.max(1, chartWidth - labelLeft - HERO_LEGEND_RIGHT_INSET_PX)
-
-      const labelMeasureContext = typeof document !== 'undefined'
-        ? document.createElement('canvas').getContext('2d')
-        : null
-      if (labelMeasureContext) {
-        labelMeasureContext.font = '500 13px ui-sans-serif, system-ui, -apple-system, "Segoe UI", sans-serif'
-      }
-
-      const preferredEntries = heroLegendSeriesWithValues.map((entry) => {
-        const clampedValue = Math.max(yBounds.min, Math.min(yBounds.max, entry.value))
-        const dotY = chartMargin.top + ((yBounds.max - clampedValue) / ySpan) * plotHeight
-        const normalizedName = entry.name.trim()
-        const measuredNameWidth = normalizedName
-          ? (labelMeasureContext?.measureText(normalizedName).width ?? normalizedName.length * 7)
-          : 0
-        const wrappedNameLineCount = Math.max(1, Math.ceil(measuredNameWidth / availableLabelWidth))
-        const labelHeight = Math.max(
-          HERO_LEGEND_MIN_HEIGHT_PX,
-          (wrappedNameLineCount * HERO_LEGEND_NAME_LINE_HEIGHT_PX) + HERO_LEGEND_VALUE_LINE_HEIGHT_PX,
-        )
-        const anchorOffset = labelHeight / 2
-        const preferredTop = dotY - anchorOffset
-        const maxTopForEntry = chartBottom - labelHeight
-
-        return {
-          ...entry,
-          dotY,
-          left: labelLeft,
-          labelHeight,
-          preferredTop: Math.max(chartTop, Math.min(maxTopForEntry, preferredTop)),
-        }
-      })
-
-      const sortedByPreferredTop = [...preferredEntries]
-        .sort((left, right) => left.preferredTop - right.preferredTop)
-
-      const stacked: Array<(typeof sortedByPreferredTop)[number] & { top: number }> = []
-      sortedByPreferredTop.forEach((entry, index) => {
-        const previousBottom = index > 0
-          ? (stacked[index - 1]!.top + stacked[index - 1]!.labelHeight)
-          : null
-        const top = previousBottom == null
-          ? entry.preferredTop
-          : Math.max(entry.preferredTop, previousBottom + HERO_LEGEND_VERTICAL_GAP_PX)
-        const maxTopForEntry = chartBottom - entry.labelHeight
-        stacked.push({ ...entry, top: Math.max(chartTop, Math.min(maxTopForEntry, top)) })
-      })
-
-      for (let index = stacked.length - 2; index >= 0; index -= 1) {
-        const entry = stacked[index]!
-        const next = stacked[index + 1]!
-        const maxTopForEntry = chartBottom - entry.labelHeight
-        const highestTopAllowedByNext = next.top - HERO_LEGEND_VERTICAL_GAP_PX - entry.labelHeight
-        entry.top = Math.max(
-          chartTop,
-          Math.min(maxTopForEntry, Math.min(entry.top, highestTopAllowedByNext)),
-        )
-      }
-
-      const topByKey = new Map(stacked.map(entry => [entry.key, entry.top] as const))
-      return preferredEntries.map(entry => ({
-        ...entry,
-        top: topByKey.get(entry.key) ?? entry.preferredTop,
-      }))
-    },
-    [
-      chartData,
-      chartHeight,
-      heroLegendRenderedWidth,
-      chartMargin.bottom,
-      chartMargin.left,
-      chartMargin.right,
-      chartMargin.top,
-      chartSeries,
-      chartWidth,
-      chartXDomain,
-      cursorSnapshot?.date,
-      heroLegendSeriesWithValues,
-      canRenderPositionedSeriesLegend,
-    ],
-  )
-
-  const legendSeriesWithValues = useMemo(
-    () => chartSeries
-      .map((seriesItem) => {
-        const hoveredValue = cursorSnapshot?.values?.[seriesItem.key]
-        const value = typeof hoveredValue === 'number' && Number.isFinite(hoveredValue)
-          ? hoveredValue
-          : latestSnapshot[seriesItem.key]
-
-        if (typeof value !== 'number' || !Number.isFinite(value)) {
-          return null
-        }
-
-        return { ...seriesItem, value }
-      })
-      .filter((entry): entry is { key: string, name: string, color: string, value: number } => entry !== null),
-    [chartSeries, cursorSnapshot, latestSnapshot],
-  )
-
-  return {
-    heroLegendRenderedWidth,
-    chartXDomain,
-    heroLegendSeriesWithValues,
-    heroLegendPositionedEntries,
-    legendSeriesWithValues,
-  }
-}
-
-function useSportsGameGraphInteractionState(defaultTimeRange: (typeof TIME_RANGES)[number]) {
-  const [cursorSnapshot, setCursorSnapshot] = useState<PredictionChartCursorSnapshot | null>(null)
-  const [activeTimeRange, setActiveTimeRange] = useState<(typeof TIME_RANGES)[number]>(defaultTimeRange)
-  const [exportDialogOpen, setExportDialogOpen] = useState(false)
-  const [embedDialogOpen, setEmbedDialogOpen] = useState(false)
-
-  return {
-    cursorSnapshot,
-    setCursorSnapshot,
-    activeTimeRange,
-    setActiveTimeRange,
-    exportDialogOpen,
-    setExportDialogOpen,
-    embedDialogOpen,
-    setEmbedDialogOpen,
-  }
-}
-
-function useSportsGameGraphTradeFlow({
-  isSportsEventHeroVariant,
-  tradeFlowSeriesByTokenId,
-}: {
-  isSportsEventHeroVariant: boolean
-  tradeFlowSeriesByTokenId: Map<string, { color: string }>
-}) {
-  const [tradeFlowItems, setTradeFlowItems] = useState<SportsTradeFlowLabelItem[]>([])
-  const tradeFlowIdRef = useRef(0)
-  const hasTradeFlowLabels = tradeFlowItems.length > 0
-
-  useOptionalMarketChannelSubscription((payload) => {
-    if (!isSportsEventHeroVariant || !payload) {
-      return
-    }
-
-    if (payload.event_type !== 'last_trade_price') {
-      return
-    }
-
-    const assetId = String(payload.asset_id ?? '')
-    if (!assetId) {
-      return
-    }
-
-    const matchedSeries = tradeFlowSeriesByTokenId.get(assetId)
-    if (!matchedSeries) {
-      return
-    }
-
-    const price = Number(payload.price)
-    const size = Number(payload.size)
-    const label = buildTradeFlowLabel(price, size)
-    if (!label) {
-      return
-    }
-
-    const createdAt = Date.now()
-    const id = String(tradeFlowIdRef.current)
-    tradeFlowIdRef.current += 1
-
-    setTradeFlowItems((previous) => {
-      const next = [...previous, { id, label, color: matchedSeries.color, createdAt }]
-      return trimTradeFlowItems(pruneTradeFlowItems(next, createdAt))
-    })
-  })
-
-  useEffect(function pruneExpiredTradeFlowItems() {
-    if (!isSportsEventHeroVariant || !hasTradeFlowLabels) {
-      return undefined
-    }
-
-    const interval = window.setInterval(() => {
-      const now = Date.now()
-      setTradeFlowItems((previous) => {
-        const next = pruneTradeFlowItems(previous, now)
-        if (next.length === previous.length) {
-          return previous
-        }
-        return next
-      })
-    }, TRADE_FLOW_CLEANUP_INTERVAL_MS)
-
-    return function clearTradeFlowPruneInterval() {
-      window.clearInterval(interval)
-    }
-  }, [hasTradeFlowLabels, isSportsEventHeroVariant])
-
-  return { tradeFlowItems, hasTradeFlowLabels }
-}
-
-export function SportsGameGraph({
-  card,
-  selectedMarketType,
-  selectedConditionId,
-  defaultTimeRange = '1W',
-  variant = 'default',
-}: {
-  card: SportsGamesCard
-  selectedMarketType: SportsGamesMarketType
-  selectedConditionId: string | null
-  defaultTimeRange?: (typeof TIME_RANGES)[number]
-  variant?: SportsGameGraphVariant
-}) {
-  const { width: windowWidth } = useWindowSize()
-  const {
-    cursorSnapshot,
-    setCursorSnapshot,
-    activeTimeRange,
-    setActiveTimeRange,
-    exportDialogOpen,
-    setExportDialogOpen,
-    embedDialogOpen,
-    setEmbedDialogOpen,
-  } = useSportsGameGraphInteractionState(defaultTimeRange)
-  const isSecondaryMarketGraph = selectedMarketType === 'spread' || selectedMarketType === 'total'
-
-  const [chartSettings, setChartSettings] = useSportsGameGraphChartSettings()
-
-  const {
-    isSportsEventHeroVariant,
-    usesPositionedSeriesLegend,
-    canRenderPositionedSeriesLegend,
-    chartHeight,
-    chartMargin,
-    chartWidth,
-  } = useSportsGameGraphChartDimensions({ windowWidth, variant })
-
-  const { graphSeriesTargets, tradeFlowSeriesByTokenId, marketTargets, chartSeries } = useSportsGameGraphSeries({
-    card,
-    selectedConditionId,
-    isSecondaryMarketGraph,
-    isSportsEventHeroVariant,
-  })
-
-  const { chartData, latestSnapshot, leadingGapStart } = useSportsGameGraphHistory({
-    card,
-    marketTargets,
-    activeTimeRange,
-    chartSeries,
-    graphSeriesTargets,
-    isSecondaryMarketGraph,
-  })
-
-  const {
-    chartXDomain,
-    heroLegendPositionedEntries,
-    legendSeriesWithValues,
-  } = useSportsGameGraphHeroLegend({
-    canRenderPositionedSeriesLegend,
-    chartSeries,
-    chartData,
-    chartWidth,
-    chartHeight,
-    chartMargin,
-    cursorSnapshot,
-    latestSnapshot,
-    usesPositionedSeriesLegend,
-  })
-
-  const { tradeFlowItems, hasTradeFlowLabels } = useSportsGameGraphTradeFlow({
-    isSportsEventHeroVariant,
-    tradeFlowSeriesByTokenId,
-  })
-
-  const legendContent = !isSecondaryMarketGraph && !usesPositionedSeriesLegend && legendSeriesWithValues.length > 0
-    ? (
-        <div className="flex min-h-5 flex-wrap items-center gap-4">
-          {legendSeriesWithValues.map(entry => (
-            <div key={entry.key} className="flex items-center gap-2">
-              <div className="size-2 shrink-0 rounded-full" style={{ backgroundColor: entry.color }} />
-              <span className="inline-flex w-fit items-center gap-2 text-xs font-medium text-muted-foreground">
-                <span>{entry.name}</span>
-                <span className={`
-                  inline-flex min-w-8 shrink-0 items-baseline justify-end text-sm font-semibold text-foreground
-                  tabular-nums
-                `}
-                >
-                  {entry.value.toFixed(0)}
-                  <span className="ml-0.5 text-sm text-foreground">%</span>
-                </span>
-              </span>
-            </div>
-          ))}
-        </div>
-      )
-    : null
-
-  if (graphSeriesTargets.length === 0) {
-    return (
-      <div className="rounded-lg border bg-secondary/30 px-3 py-6 text-sm text-muted-foreground">
-        Graph is unavailable for this game.
-      </div>
-    )
-  }
-
-  return (
-    <>
-      <div style={usesPositionedSeriesLegend ? { minHeight: `${chartHeight + 56}px` } : undefined}>
-        <div className="relative">
-          <PredictionChart
-            data={chartData}
-            series={chartSeries}
-            width={chartWidth}
-            height={chartHeight}
-            margin={chartMargin}
-            xDomain={chartXDomain}
-            dataSignature={`${card.id}:${chartSeries.map(series => series.key).join(',')}:${activeTimeRange}`}
-            onCursorDataChange={setCursorSnapshot}
-            xAxisTickCount={3}
-            yAxis={undefined}
-            legendContent={legendContent}
-            showLegend={!isSecondaryMarketGraph && !usesPositionedSeriesLegend}
-            showTooltipSeriesLabels={!usesPositionedSeriesLegend}
-            disableCursorSplit={false}
-            clampCursorToDataExtent={usesPositionedSeriesLegend}
-            markerOuterRadius={usesPositionedSeriesLegend ? 10 : undefined}
-            markerInnerRadius={usesPositionedSeriesLegend ? 4.2 : undefined}
-            markerPulseStyle={usesPositionedSeriesLegend ? 'ring' : undefined}
-            lineCurve="monotoneX"
-            tooltipValueFormatter={value => `${Math.round(value)}%`}
-            autoscale={chartSettings.autoscale}
-            showXAxis={chartSettings.xAxis}
-            showYAxis={chartSettings.yAxis}
-            showHorizontalGrid={chartSettings.horizontalGrid}
-            showVerticalGrid={chartSettings.verticalGrid}
-            showAnnotations={chartSettings.annotations}
-            leadingGapStart={leadingGapStart}
-          />
-
-          {canRenderPositionedSeriesLegend && heroLegendPositionedEntries.length > 0 && (
-            <div className="pointer-events-none absolute inset-0 overflow-hidden">
-              {heroLegendPositionedEntries.map(entry => (
-                <div
-                  key={entry.key}
-                  className="absolute"
-                  style={{
-                    top: `${entry.top}px`,
-                    left: `${entry.left}px`,
-                  }}
-                >
-                  <p
-                    className={cn(
-                      'text-[13px] leading-snug font-medium tracking-tight',
-                      (windowWidth ?? 1024) >= 768 && 'truncate',
-                    )}
-                    style={{ color: entry.color }}
-                  >
-                    {entry.name}
-                  </p>
-                  <p
-                    className="text-2xl/tight font-semibold tabular-nums"
-                    style={{ color: entry.color }}
-                  >
-                    {`${Math.round(entry.value)}%`}
-                  </p>
-                </div>
-              ))}
-            </div>
-          )}
-
-          {isSportsEventHeroVariant && hasTradeFlowLabels && (
-            <div className={`
-              pointer-events-none absolute bottom-6 left-4 flex flex-col gap-1 text-sm font-semibold tabular-nums
-            `}
-            >
-              {tradeFlowItems.map(item => (
-                <span
-                  key={item.id}
-                  className="animate-trade-flow-rise"
-                  style={{
-                    ...tradeFlowTextStrokeStyle,
-                    color: item.color,
-                  }}
-                >
-                  +
-                  {item.label}
-                </span>
-              ))}
-            </div>
-          )}
-        </div>
-
-        <div className="mt-2 flex items-center justify-end pb-2">
-          <EventChartControls
-            timeRanges={TIME_RANGES}
-            activeTimeRange={activeTimeRange}
-            onTimeRangeChange={setActiveTimeRange}
-            showOutcomeSwitch={false}
-            oppositeOutcomeLabel=""
-            onShuffle={() => {}}
-            settings={chartSettings}
-            onSettingsChange={setChartSettings}
-            onExportData={() => setExportDialogOpen(true)}
-            onEmbed={() => setEmbedDialogOpen(true)}
-          />
-        </div>
-      </div>
-
-      <EventChartExportDialog
-        open={exportDialogOpen}
-        onOpenChange={setExportDialogOpen}
-        eventCreatedAt={card.eventCreatedAt}
-        markets={card.detailMarkets}
-        isMultiMarket={card.detailMarkets.length > 1}
-      />
-      <EventChartEmbedDialog
-        open={embedDialogOpen}
-        onOpenChange={setEmbedDialogOpen}
-        markets={card.detailMarkets}
-        initialMarketId={selectedConditionId}
-      />
-    </>
-  )
-}
-
-function resolveTeamShortLabel(name: string | null | undefined, abbreviation: string | null | undefined) {
-  const normalizedAbbreviation = abbreviation
-    ?.trim()
-    .replace(/[^a-z0-9]/gi, '')
-    .toUpperCase()
-  if (normalizedAbbreviation) {
-    return normalizedAbbreviation
-  }
-
-  const compactName = name
-    ?.trim()
-    .replace(/[^a-z0-9]/gi, '')
-    .toUpperCase()
-  if (!compactName) {
-    return null
-  }
-
-  return compactName.slice(0, 3)
-}
-
-const FRANCHISE_MULTI_WORD_NICKNAME_PREFIXES = new Set([
-  'blue',
-  'golden',
-  'maple',
-  'red',
-  'trail',
-  'white',
-])
-
-function resolveFranchiseTradeHeaderTeamLabel(
-  name: string | null | undefined,
-  abbreviation: string | null | undefined,
-) {
-  const trimmedName = name?.trim()
-  if (!trimmedName) {
-    return resolveTeamShortLabel(name, abbreviation)
-  }
-
-  const nameTokens = trimmedName.split(/\s+/).filter(Boolean)
-  if (nameTokens.length <= 1) {
-    return trimmedName
-  }
-
-  const lastToken = nameTokens.at(-1)
-  const secondToLastToken = nameTokens.at(-2)
-  if (!lastToken) {
-    return resolveTeamShortLabel(name, abbreviation)
-  }
-
-  if (
-    secondToLastToken
-    && FRANCHISE_MULTI_WORD_NICKNAME_PREFIXES.has(normalizeComparableText(secondToLastToken))
-  ) {
-    return `${secondToLastToken} ${lastToken}`
-  }
-
-  return lastToken
-}
-
-function resolveEsportsTradeHeaderTeamLabel(
-  name: string | null | undefined,
-  abbreviation: string | null | undefined,
-) {
-  const trimmedAbbreviation = abbreviation?.trim()
-  if (trimmedAbbreviation) {
-    return trimmedAbbreviation
-      .toUpperCase()
-      .replace(/[_-]+/g, ' ')
-      .replace(/([A-Z]+)(\d+)$/u, '$1 $2')
-      .replace(/\s+/g, ' ')
-      .trim()
-  }
-
-  return resolveTeamShortLabel(name, abbreviation)
-}
-
-const COMPACT_COMBAT_TRADE_HEADER_SPORT_SLUGS = new Set([
-  'boxing',
-  'chess',
-  'mma',
-  'ufc',
-  'zuffa',
-])
-
-const COMPACT_FRANCHISE_TRADE_HEADER_SPORT_SLUGS = new Set([
-  'mlb',
-  'nba',
-  'nfl',
-  'nhl',
-  'wnba',
-])
-
-function resolveTradeHeaderSportSlugCandidates(card: SportsGamesCard) {
-  return [
-    card.event.sports_sport_slug,
-    card.event.sports_series_slug,
-  ]
-    .map(value => normalizeComparableText(value))
-    .filter((value): value is string => Boolean(value))
-}
-
-function isCompactCricketTradeHeaderSportSlug(slug: string) {
-  return slug === 'cricket' || slug === 'crint' || slug.startsWith('cric')
-}
-
-function shouldUseFranchiseTradeHeaderTeamLabels(sportSlugs: string[]) {
-  return sportSlugs.some(slug => COMPACT_FRANCHISE_TRADE_HEADER_SPORT_SLUGS.has(slug))
-}
-
-function hasDrawMoneylineOption(card: SportsGamesCard) {
-  return card.buttons.some(button =>
-    button.marketType === 'moneyline' && button.tone === 'draw',
-  )
-}
-
-function resolveCompactTradeHeaderTitle(
-  card: SportsGamesCard,
-  resolveTeamLabel: (
-    name: string | null | undefined,
-    abbreviation: string | null | undefined,
-  ) => string | null,
-) {
-  const team1 = card.teams[0] ?? null
-  const team2 = card.teams[1] ?? null
-  const leftLabel = resolveTeamLabel(team1?.name, team1?.abbreviation)
-  const rightLabel = resolveTeamLabel(team2?.name, team2?.abbreviation)
-
-  if (!leftLabel || !rightLabel) {
-    return null
-  }
-
-  return `${leftLabel} vs ${rightLabel}`
-}
-
-function shouldUseCompactTradeHeaderTitle(card: SportsGamesCard, vertical: SportsVertical | null) {
-  if (vertical === 'esports') {
-    return true
-  }
-
-  if (hasDrawMoneylineOption(card)) {
-    return true
-  }
-
-  const sportSlugs = resolveTradeHeaderSportSlugCandidates(card)
-
-  return sportSlugs.some(slug =>
-    COMPACT_COMBAT_TRADE_HEADER_SPORT_SLUGS.has(slug)
-    || COMPACT_FRANCHISE_TRADE_HEADER_SPORT_SLUGS.has(slug)
-    || isCompactCricketTradeHeaderSportSlug(slug))
-}
-
-function escapeRegExp(value: string) {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-}
-
-function abbreviatePositionMarketLabel(label: string, teams: SportsGamesCard['teams']) {
-  const trimmedLabel = label.trim()
-  if (!trimmedLabel) {
-    return ''
-  }
-
-  let nextLabel = trimmedLabel
-  const replacements = teams
-    .map(team => ({
-      teamName: team.name.trim(),
-      shortLabel: resolveTeamShortLabel(team.name, team.abbreviation),
-    }))
-    .filter(({ teamName, shortLabel }) => teamName.length > 0 && Boolean(shortLabel))
-    .sort((a, b) => b.teamName.length - a.teamName.length)
-
-  for (const { teamName, shortLabel } of replacements) {
-    nextLabel = nextLabel.replace(new RegExp(escapeRegExp(teamName), 'gi'), shortLabel!)
-  }
-
-  return nextLabel.replace(/\s+/g, ' ').trim().toUpperCase()
-}
-
-function resolveTradeHeaderTitle({
-  card,
-  selectedButton,
-  selectedMarket,
-  marketType,
-}: {
-  card: SportsGamesCard
-  selectedButton: SportsGamesButton
-  selectedMarket: Market | null
-  marketType: SportsGamesMarketType
-}) {
-  const normalizedMarketType = normalizeComparableText(selectedMarket?.sports_market_type)
-  if (normalizedMarketType.includes('exact score')) {
-    const descriptor = resolveMarketDescriptor(selectedMarket)
-    return descriptor ? `Exact Score: ${descriptor}` : 'Exact Score'
-  }
-
-  if (marketType === 'btts') {
-    return 'Both Teams to Score?'
-  }
-
-  if (marketType === 'total') {
-    return 'Over vs Under'
-  }
-
-  const team1 = card.teams[0] ?? null
-  const team2 = card.teams[1] ?? null
-  const vertical = resolveSportsVerticalFromTags({
-    tags: card.event.tags,
-    mainTag: card.event.main_tag,
-  })
-  if (shouldUseCompactTradeHeaderTitle(card, vertical)) {
-    const sportSlugs = resolveTradeHeaderSportSlugCandidates(card)
-    const compactTitle = resolveCompactTradeHeaderTitle(
-      card,
-      vertical === 'esports'
-        ? resolveEsportsTradeHeaderTeamLabel
-        : shouldUseFranchiseTradeHeaderTeamLabels(sportSlugs)
-          ? resolveFranchiseTradeHeaderTeamLabel
-          : resolveTeamShortLabel,
-    )
-    if (compactTitle) {
-      return compactTitle
-    }
-  }
-
-  const fullMatchupTitle = card.title?.trim()
-    || [team1?.name?.trim(), team2?.name?.trim()].filter(Boolean).join(' vs ')
-
-  if (fullMatchupTitle) {
-    return fullMatchupTitle
-  }
-
-  return selectedButton.label.trim().toUpperCase() || card.title
-}
-
-function resolveHexToRgbComponents(value: string) {
-  const hex = value.replace('#', '')
-  const expandedHex = hex.length === 3
-    ? hex.split('').map(char => `${char}${char}`).join('')
-    : hex
-
-  const red = Number.parseInt(expandedHex.slice(0, 2), 16)
-  const green = Number.parseInt(expandedHex.slice(2, 4), 16)
-  const blue = Number.parseInt(expandedHex.slice(4, 6), 16)
-  if ([red, green, blue].some(component => Number.isNaN(component))) {
-    return null
-  }
-
-  return `${red} ${green} ${blue}`
-}
-
-function resolveTradeHeaderBadgeAccent(button: SportsGamesButton) {
-  const normalizedTeamColor = normalizeHexColor(button.color)
-  if (
-    (button.tone === 'team1' || button.tone === 'team2')
-    && normalizedTeamColor
-  ) {
-    const rgbComponents = resolveHexToRgbComponents(normalizedTeamColor)
-    return {
-      className: 'dark:mix-blend-plus-lighter',
-      style: {
-        color: normalizedTeamColor,
-        backgroundColor: rgbComponents ? `rgb(${rgbComponents} / 0.10)` : undefined,
-      } as CSSProperties,
-    }
-  }
-
-  if (button.tone === 'over') {
-    return {
-      className: 'bg-yes/10 text-yes',
-      style: undefined,
-    }
-  }
-
-  if (button.tone === 'under') {
-    return {
-      className: 'bg-no/10 text-no',
-      style: undefined,
-    }
-  }
-
-  return {
-    className: 'bg-muted/60 text-muted-foreground',
-    style: undefined,
-  }
-}
-
-function normalizeComparableToken(value: string | null | undefined) {
-  return value
-    ?.normalize('NFKD')
-    .replace(/[\u0300-\u036F]/g, '')
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '')
-    .trim()
-    ?? ''
-}
-
-function resolveTeamByTone(card: SportsGamesCard, tone: SportsGamesButton['tone']) {
-  if (tone === 'team1') {
-    return card.teams[0] ?? null
-  }
-  if (tone === 'team2') {
-    return card.teams[1] ?? null
-  }
-  return null
-}
-
-function resolveLeadingSpreadTeam(card: SportsGamesCard, button: SportsGamesButton) {
-  const firstToken = button.label.split(/\s+/)[0] ?? ''
-  const normalizedFirstToken = normalizeComparableToken(firstToken)
-  if (normalizedFirstToken) {
-    const matchedTeam = card.teams.find((team) => {
-      const abbreviationToken = normalizeComparableToken(team.abbreviation)
-      if (abbreviationToken && abbreviationToken === normalizedFirstToken) {
-        return true
-      }
-
-      const nameToken = normalizeComparableToken(team.name)
-      return Boolean(nameToken && nameToken.startsWith(normalizedFirstToken))
-    })
-
-    if (matchedTeam) {
-      return matchedTeam
-    }
-  }
-
-  return resolveTeamByTone(card, button.tone)
-}
-
-function TeamLogoBadge({
-  card,
-  button,
-}: {
-  card: SportsGamesCard
-  button: SportsGamesButton
-}) {
-  const useCroppedTeamLogo = shouldUseCroppedSportsTeamLogo(card.event.sports_sport_slug)
-  const team = button.marketType === 'spread'
-    ? resolveLeadingSpreadTeam(card, button)
-    : resolveTeamByTone(card, button.tone)
-  const fallbackInitial = team?.abbreviation?.slice(0, 1).toUpperCase()
-    || team?.name?.slice(0, 1).toUpperCase()
-    || '?'
-
-  return (
-    <div
-      className={cn(
-        'flex items-center justify-center',
-        useCroppedTeamLogo ? 'relative size-11 overflow-hidden rounded-lg' : 'size-11',
-      )}
-    >
-      {team?.logoUrl
-        ? (
-            useCroppedTeamLogo
-              ? (
-                  <Image
-                    src={team.logoUrl}
-                    alt={`${team.name} logo`}
-                    fill
-                    sizes="44px"
-                    className="scale-[1.12] object-cover object-center"
-                  />
-                )
-              : (
-                  <Image
-                    src={team.logoUrl}
-                    alt={`${team.name} logo`}
-                    width={44}
-                    height={44}
-                    sizes="44px"
-                    className="h-[92%] w-[92%] object-contain object-center"
-                  />
-                )
-          )
-        : (
-            <div
-              className={cn(
-                'flex size-full items-center justify-center text-sm font-semibold text-muted-foreground',
-                useCroppedTeamLogo && 'rounded-lg border border-border/40 bg-secondary',
-              )}
-            >
-              {fallbackInitial}
-            </div>
-          )}
-    </div>
-  )
-}
-
-function DrawBadge() {
-  return (
-    <div className="flex size-11 items-center justify-center rounded-lg bg-secondary text-muted-foreground shadow-sm">
-      <EqualIcon className="size-5.5" />
-    </div>
-  )
-}
-
-function TotalBadge({ button }: { button: SportsGamesButton }) {
-  const isOverActive = button.tone === 'over'
-  const isUnderActive = button.tone === 'under'
-
-  return (
-    <div
-      className={`
-        relative inline-flex size-11 items-center justify-center overflow-hidden rounded-lg text-white shadow-sm
-      `}
-    >
-      <span
-        className={cn(
-          'absolute inset-0 bg-yes transition-opacity [clip-path:polygon(0_0,100%_0,0_100%)]',
-          !isOverActive && 'opacity-25',
-        )}
-      />
-      <span
-        className={cn(
-          'absolute inset-0 bg-no transition-opacity [clip-path:polygon(100%_0,100%_100%,0_100%)]',
-          !isUnderActive && 'opacity-25',
-        )}
-      />
-      <span className={cn(
-        'absolute top-2 left-2 z-10 text-[11px] leading-none font-bold tracking-wide',
-        !isOverActive && 'opacity-35',
-      )}
-      >
-        O
-      </span>
-      <span className={cn(
-        'absolute right-2 bottom-2 z-10 text-[11px] leading-none font-bold tracking-wide',
-        !isUnderActive && 'opacity-35',
-      )}
-      >
-        U
-      </span>
-    </div>
-  )
-}
-
-function BttsBadge({ button }: { button: SportsGamesButton }) {
-  const isYesActive = button.tone !== 'under'
-  const isNoActive = button.tone === 'under'
-
-  return (
-    <div
-      className={`
-        relative inline-flex size-11 items-center justify-center overflow-hidden rounded-lg text-white shadow-sm
-      `}
-    >
-      <span
-        className={cn(
-          'absolute inset-0 bg-yes transition-opacity [clip-path:polygon(0_0,100%_0,0_100%)]',
-          !isYesActive && 'opacity-25',
-        )}
-      />
-      <span
-        className={cn(
-          'absolute inset-0 bg-no transition-opacity [clip-path:polygon(100%_0,100%_100%,0_100%)]',
-          !isNoActive && 'opacity-25',
-        )}
-      />
-      <span className={cn(
-        'absolute top-2 left-2 z-10 text-[11px] leading-none font-bold tracking-wide',
-        !isYesActive && 'opacity-35',
-      )}
-      >
-        Y
-      </span>
-      <span className={cn(
-        'absolute right-2 bottom-2 z-10 text-[11px] leading-none font-bold tracking-wide',
-        !isNoActive && 'opacity-35',
-      )}
-      >
-        N
-      </span>
-    </div>
-  )
-}
-
-export function SportsOrderPanelMarketInfo({
-  card,
-  selectedButton,
-  selectedOutcome,
-  marketType,
-}: {
-  card: SportsGamesCard
-  selectedButton: SportsGamesButton
-  selectedOutcome: Outcome | null
-  marketType: SportsGamesMarketType
-}) {
-  const selectedMarket = resolveSelectedMarket(card, selectedButton.key)
-  const badgeLabel = resolveSelectedTradeLabel(card, selectedButton, selectedOutcome)
-  const headerTitle = resolveTradeHeaderTitle({
-    card,
-    selectedButton,
-    selectedMarket,
-    marketType,
-  })
-  const badgeAccent = resolveTradeHeaderBadgeAccent(selectedButton)
-  const isExactScoreTrade = normalizeComparableText(selectedMarket?.sports_market_type).includes('exact score')
-  let marketIcon: React.ReactNode = null
-  if (!isExactScoreTrade) {
-    if (marketType === 'total') {
-      marketIcon = <TotalBadge button={selectedButton} />
-    }
-    else if (marketType === 'btts') {
-      marketIcon = <BttsBadge button={selectedButton} />
-    }
-    else if (selectedButton.tone === 'draw') {
-      marketIcon = <DrawBadge />
-    }
-    else {
-      marketIcon = <TeamLogoBadge card={card} button={selectedButton} />
-    }
-  }
-
-  return (
-    <div className="mb-4">
-      <div className={cn('flex items-start', marketIcon && 'gap-3')}>
-        {marketIcon && (
-          <div className="shrink-0">
-            {marketIcon}
-          </div>
-        )}
-
-        <div className="min-w-0">
-          <p className="line-clamp-2 text-base/tight font-bold text-foreground">
-            {headerTitle}
-          </p>
-          <span
-            className={cn(
-              'mt-1.5 inline-flex items-center rounded-sm px-2.5 py-1 text-xs font-semibold',
-              badgeAccent.className,
-            )}
-            style={badgeAccent.style}
-          >
-            {badgeLabel}
-          </span>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-interface SportsGameDetailsPanelProps {
-  card: SportsGamesCard
-  activeDetailsTab: DetailsTab
-  selectedButtonKey: string | null
-  showBottomContent: boolean
-  defaultGraphTimeRange?: (typeof TIME_RANGES)[number]
-  allowedConditionIds?: Set<string> | null
-  positionsTitle?: string
-  showAboutTab?: boolean
-  aboutEvent?: SportsGamesCard['event'] | null
-  rulesEvent?: SportsGamesCard['event'] | null
-  showRedeemInPositions?: boolean
-  onOpenRedeemForCondition?: ((conditionId: string) => void) | null
-  oddsFormat?: OddsFormat
-  marketContextEnabled?: boolean
-  onChangeTab: (tab: DetailsTab) => void
-  onSelectButton: (
-    buttonKey: string,
-    options?: { panelMode?: 'full' | 'partial' | 'preserve' },
-  ) => void
-}
-
-function useSportsGameDetailsPanelOrderStore() {
-  const orderMarketConditionId = useOrder(state => state.market?.condition_id ?? null)
-  const orderOutcomeIndex = useOrder(state => state.outcome?.outcome_index ?? null)
-  const setOrderOutcome = useOrder(state => state.setOutcome)
-  const setOrderMarket = useOrder(state => state.setMarket)
-  const setOrderType = useOrder(state => state.setType)
-  const setOrderSide = useOrder(state => state.setSide)
-  const setOrderAmount = useOrder(state => state.setAmount)
-  const setIsMobileOrderPanelOpen = useOrder(state => state.setIsMobileOrderPanelOpen)
-
-  return {
-    orderMarketConditionId,
-    orderOutcomeIndex,
-    setOrderOutcome,
-    setOrderMarket,
-    setOrderType,
-    setOrderSide,
-    setOrderAmount,
-    setIsMobileOrderPanelOpen,
-  }
-}
-
-function useSportsGameDetailsPanelLocalState() {
-  const [cashOutPayload, setCashOutPayload] = useState<SportsCashOutModalPayload | null>(null)
-  const [isPositionsExpanded, setIsPositionsExpanded] = useState(false)
-  const [convertTagKey, setConvertTagKey] = useState<string | null>(null)
-
-  return {
-    cashOutPayload,
-    setCashOutPayload,
-    isPositionsExpanded,
-    setIsPositionsExpanded,
-    convertTagKey,
-    setConvertTagKey,
-  }
-}
-
-function useSportsCardDerivations(card: SportsGamesCard) {
-  const cardMarketByConditionId = useMemo(
-    () => new Map(card.detailMarkets.map(market => [market.condition_id, market] as const)),
-    [card.detailMarkets],
-  )
-
-  const cardButtonsByConditionAndOutcome = useMemo(() => {
-    const map = new Map<string, SportsGamesButton>()
-    card.buttons.forEach((button) => {
-      map.set(`${button.conditionId}:${button.outcomeIndex}`, button)
-    })
-    return map
-  }, [card.buttons])
-
-  const cardFirstButtonByCondition = useMemo(() => {
-    const map = new Map<string, SportsGamesButton>()
-    card.buttons.forEach((button) => {
-      if (!map.has(button.conditionId)) {
-        map.set(button.conditionId, button)
-      }
-    })
-    return map
-  }, [card.buttons])
-
-  const moneylineConditionIds = useMemo(
-    () => new Set(
-      card.buttons
-        .filter(button => button.marketType === 'moneyline')
-        .map(button => button.conditionId),
-    ),
-    [card.buttons],
-  )
-
-  const isNegRiskEnabled = useMemo(() => {
-    return Boolean(
-      card.event.neg_risk
-      || card.event.neg_risk_augmented
-      || card.event.neg_risk_market_id
-      || card.detailMarkets.some(market => market.neg_risk || market.neg_risk_market_id),
-    )
-  }, [card.detailMarkets, card.event.neg_risk, card.event.neg_risk_augmented, card.event.neg_risk_market_id])
-
-  return {
-    cardMarketByConditionId,
-    cardButtonsByConditionAndOutcome,
-    cardFirstButtonByCondition,
-    moneylineConditionIds,
-    isNegRiskEnabled,
-  }
-}
-
-function useSportsCardUserPositionsQuery({
-  ownerAddress,
-  cardId,
-  showBottomContent,
-}: {
-  ownerAddress: string | null
-  cardId: string
-  showBottomContent: boolean
-}) {
-  const { data: userPositions } = useQuery<UserPosition[]>({
-    queryKey: ['sports-card-user-positions', ownerAddress, cardId],
-    enabled: Boolean(ownerAddress),
-    staleTime: 1000 * 30,
-    gcTime: 1000 * 60 * 10,
-    refetchInterval: ownerAddress ? (showBottomContent ? 15_000 : false) : false,
-    refetchIntervalInBackground: showBottomContent,
-    queryFn: ({ signal }) => fetchUserPositionsForMarket({
-      pageParam: 0,
-      userAddress: ownerAddress!,
-      status: 'active',
-      signal,
-    }),
-  })
-
-  return userPositions
-}
-
-function useSportsPositionTags({
-  ownerAddress,
-  userPositions,
-  allowedConditionIds,
-  card,
-  cardMarketByConditionId,
-  cardButtonsByConditionAndOutcome,
-  cardFirstButtonByCondition,
-}: {
-  ownerAddress: string | null
-  userPositions: UserPosition[] | undefined
-  allowedConditionIds: Set<string> | null
-  card: SportsGamesCard
-  cardMarketByConditionId: Map<string, Market>
-  cardButtonsByConditionAndOutcome: Map<string, SportsGamesButton>
-  cardFirstButtonByCondition: Map<string, SportsGamesButton>
-}) {
-  const positionTags = useMemo<SportsPositionTag[]>(() => {
-    if (!ownerAddress || !userPositions?.length) {
-      return []
-    }
-
-    const aggregated = new Map<string, {
-      conditionId: string
-      outcomeIndex: typeof OUTCOME_INDEX.YES | typeof OUTCOME_INDEX.NO
-      market: Market
-      outcome: Outcome
-      button: SportsGamesButton | null
-      marketTypeLabel: 'Moneyline' | 'Spread' | 'Total' | 'Both Teams to Score' | 'Market'
-      marketLabel: string
-      outcomeLabel: string
-      shares: number
-      totalCost: number | null
-      currentValue: number
-      realizedPnl: number
-      latestActivityAtMs: number
-    }>()
-
-    userPositions.forEach((position) => {
-      const conditionId = position.market?.condition_id
-      if (!conditionId) {
-        return
-      }
-
-      if (allowedConditionIds && !allowedConditionIds.has(conditionId)) {
-        return
-      }
-
-      const market = cardMarketByConditionId.get(conditionId)
-      if (!market) {
-        return
-      }
-
-      const shares = resolvePositionShares(position)
-      if (!(shares > 0)) {
-        return
-      }
-
-      const explicitOutcomeIndex = typeof position.outcome_index === 'number'
-        ? position.outcome_index
-        : undefined
-      const normalizedOutcomeText = position.outcome_text?.trim().toLowerCase()
-      const resolvedOutcomeIndex = explicitOutcomeIndex ?? normalizedOutcomeText === 'no'
-        ? OUTCOME_INDEX.NO
-        : OUTCOME_INDEX.YES
-
-      if (resolvedOutcomeIndex !== OUTCOME_INDEX.YES && resolvedOutcomeIndex !== OUTCOME_INDEX.NO) {
-        return
-      }
-
-      const outcome = market.outcomes.find(item => item.outcome_index === resolvedOutcomeIndex)
-      if (!outcome) {
-        return
-      }
-
-      const button = cardButtonsByConditionAndOutcome.get(`${conditionId}:${resolvedOutcomeIndex}`)
-        ?? cardFirstButtonByCondition.get(conditionId)
-        ?? null
-      const fallbackMarketLabel = market.sports_group_item_title?.trim()
-        || market.short_title?.trim()
-        || market.title
-      const rawMarketLabel = button?.marketType === 'binary'
-        ? fallbackMarketLabel
-        : button?.label?.trim()
-          || outcome.outcome_text?.trim()
-          || fallbackMarketLabel
-      const marketLabel = abbreviatePositionMarketLabel(rawMarketLabel, card.teams)
-        || abbreviatePositionMarketLabel(fallbackMarketLabel, card.teams)
-      const outcomeLabel = resolvedOutcomeIndex === OUTCOME_INDEX.NO ? 'NO' : 'YES'
-      const avgPrice = normalizePositionPrice(position.avgPrice)
-        ?? normalizePositionPrice(Number(fromMicro(String(position.average_position ?? 0), 6)))
-      const normalizedAvgPrice = Number.isFinite(avgPrice) ? avgPrice : null
-      const costValue = resolvePositionCostValue(position, shares, normalizedAvgPrice)
-      const normalizedMarketPrice = normalizePositionPrice(outcome.buy_price)
-      const currentValue = resolvePositionCurrentValue(position, shares, normalizedAvgPrice, normalizedMarketPrice)
-      const rawRealizedPnl = toFiniteNumber(position.realizedPnl)
-        ?? toFiniteNumber(position.cashPnl)
-        ?? 0
-      const realizedPnl = normalizePositionPnlValue(rawRealizedPnl, costValue)
-      const activityMs = Date.parse(position.last_activity_at)
-      const normalizedActivityMs = Number.isFinite(activityMs) ? activityMs : 0
-      const key = `${conditionId}:${resolvedOutcomeIndex}`
-      const existing = aggregated.get(key)
-
-      if (!existing) {
-        aggregated.set(key, {
-          conditionId,
-          outcomeIndex: resolvedOutcomeIndex,
-          market,
-          outcome,
-          button,
-          marketTypeLabel: resolveMarketTypeLabel(button, market),
-          marketLabel,
-          outcomeLabel,
-          shares,
-          totalCost: typeof costValue === 'number' ? costValue : null,
-          currentValue,
-          realizedPnl,
-          latestActivityAtMs: normalizedActivityMs,
-        })
-        return
-      }
-
-      existing.shares += shares
-      existing.currentValue += currentValue
-      existing.realizedPnl += realizedPnl
-      existing.latestActivityAtMs = Math.max(existing.latestActivityAtMs, normalizedActivityMs)
-      if (typeof costValue === 'number') {
-        existing.totalCost = (existing.totalCost ?? 0) + costValue
-      }
-    })
-
-    return Array.from(aggregated.values())
-      .map((item) => {
-        const avgPriceCents = item.shares > 0 && typeof item.totalCost === 'number'
-          ? (item.totalCost / item.shares) * 100
-          : null
-        const summaryLabel = item.marketTypeLabel === 'Moneyline' || item.marketTypeLabel === 'Market'
-          ? `${item.marketLabel} ${item.outcomeLabel}`.trim()
-          : item.marketLabel.trim()
-
-        return {
-          key: `${item.conditionId}:${item.outcomeIndex}`,
-          conditionId: item.conditionId,
-          outcomeIndex: item.outcomeIndex,
-          marketTypeLabel: item.marketTypeLabel,
-          marketLabel: item.marketLabel,
-          outcomeLabel: item.outcomeLabel,
-          summaryLabel,
-          shares: item.shares,
-          avgPriceCents,
-          totalCost: item.totalCost,
-          currentValue: item.currentValue,
-          realizedPnl: item.realizedPnl,
-          market: item.market,
-          outcome: item.outcome,
-          button: item.button,
-          latestActivityAtMs: item.latestActivityAtMs,
-        }
-      })
-      .sort((a, b) => b.latestActivityAtMs - a.latestActivityAtMs)
-  }, [
-    allowedConditionIds,
-    card.teams,
-    cardButtonsByConditionAndOutcome,
-    cardFirstButtonByCondition,
-    cardMarketByConditionId,
-    ownerAddress,
-    userPositions,
-  ])
-
-  const visiblePositionTags = useMemo(
-    () => positionTags.slice(0, 3),
-    [positionTags],
-  )
-
-  const hiddenPositionTagsCount = useMemo(
-    () => Math.max(0, positionTags.length - visiblePositionTags.length),
-    [positionTags.length, visiblePositionTags.length],
-  )
-
-  return { positionTags, visiblePositionTags, hiddenPositionTagsCount }
-}
-
-function useSportsConvertDialog({
-  convertTagKey,
-  positionTags,
-  card,
-  allowedConditionIds,
-}: {
-  convertTagKey: string | null
-  positionTags: SportsPositionTag[]
-  card: SportsGamesCard
-  allowedConditionIds: Set<string> | null
-}) {
-  const activeConvertTagKey = useMemo(
-    () => (convertTagKey && positionTags.some(tag => tag.key === convertTagKey) ? convertTagKey : null),
-    [convertTagKey, positionTags],
-  )
-
-  const convertDialogTag = useMemo(
-    () => (activeConvertTagKey ? positionTags.find(tag => tag.key === activeConvertTagKey) ?? null : null),
-    [activeConvertTagKey, positionTags],
-  )
-
-  const convertDialogOptions = useMemo(() => {
-    if (!convertDialogTag) {
-      return []
-    }
-
-    return [{
-      id: convertDialogTag.key,
-      conditionId: convertDialogTag.conditionId,
-      label: convertDialogTag.market.short_title || convertDialogTag.market.title,
-      shares: convertDialogTag.shares,
-    }]
-  }, [convertDialogTag])
-
-  const convertDialogOutcomes = useMemo(() => {
-    const seenConditionIds = new Set<string>()
-    return card.detailMarkets
-      .filter((market) => {
-        if (!market.condition_id || seenConditionIds.has(market.condition_id)) {
-          return false
-        }
-        if (allowedConditionIds && !allowedConditionIds.has(market.condition_id)) {
-          return false
-        }
-        seenConditionIds.add(market.condition_id)
-        return true
-      })
-      .map(market => ({
-        conditionId: market.condition_id,
-        questionId: market.question_id,
-        label: market.short_title || market.title,
-        iconUrl: market.icon_url,
-      }))
-  }, [allowedConditionIds, card.detailMarkets])
-
-  return { convertDialogTag, convertDialogOptions, convertDialogOutcomes }
-}
-
-function useSportsSelectedMarketDerivations({
-  card,
-  selectedButtonKey,
-  orderMarketConditionId,
-  orderOutcomeIndex,
-}: {
-  card: SportsGamesCard
-  selectedButtonKey: string | null
-  orderMarketConditionId: string | null
-  orderOutcomeIndex: number | null
-}) {
-  const selectedButton = useMemo(
-    () => resolveSelectedButton(card, selectedButtonKey),
-    [card, selectedButtonKey],
-  )
-
-  const selectedMarket = useMemo(
-    () => resolveSelectedMarket(card, selectedButtonKey),
-    [card, selectedButtonKey],
-  )
-
-  const selectedOutcome = useMemo(() => {
-    if (!selectedMarket) {
-      return null
-    }
-
-    if (
-      orderMarketConditionId === selectedMarket.condition_id
-      && (orderOutcomeIndex === OUTCOME_INDEX.YES || orderOutcomeIndex === OUTCOME_INDEX.NO)
-    ) {
-      const syncedOutcome = selectedMarket.outcomes.find(
-        outcome => outcome.outcome_index === orderOutcomeIndex,
-      )
-      if (syncedOutcome) {
-        return syncedOutcome
-      }
-    }
-
-    return resolveSelectedOutcome(selectedMarket, selectedButton)
-  }, [orderMarketConditionId, orderOutcomeIndex, selectedButton, selectedMarket])
-
-  const selectedLinePickerMarketType = useMemo<LinePickerMarketType | null>(() => {
-    if (!selectedButton) {
-      return null
-    }
-    return (selectedButton.marketType === 'spread' || selectedButton.marketType === 'total')
-      ? selectedButton.marketType
-      : null
-  }, [selectedButton])
-
-  const nextOutcome = useMemo(() => {
-    if (!selectedMarket || !selectedOutcome) {
-      return null
-    }
-
-    return selectedMarket.outcomes.find(
-      outcome => outcome.outcome_index !== selectedOutcome.outcome_index,
-    ) ?? null
-  }, [selectedMarket, selectedOutcome])
-
-  const nextButton = useMemo(() => {
-    if (!selectedMarket || !nextOutcome) {
-      return null
-    }
-
-    return card.buttons.find(
-      button => button.conditionId === selectedMarket.condition_id
-        && button.outcomeIndex === nextOutcome.outcome_index,
-    ) ?? null
-  }, [card.buttons, nextOutcome, selectedMarket])
-
-  const tradeSelectionLabel = useMemo(
-    () => resolveSelectedTradeLabel(card, selectedButton, selectedOutcome),
-    [card, selectedButton, selectedOutcome],
-  )
-
-  const switchTooltip = useMemo(() => {
-    return resolveSwitchTooltip(selectedMarket, nextOutcome)
-  }, [nextOutcome, selectedMarket])
-
-  const selectedMarketTokenIds = useMemo(() => {
-    if (!selectedMarket) {
-      return []
-    }
-
-    return selectedMarket.outcomes
-      .map(outcome => outcome.token_id)
-      .filter((tokenId): tokenId is string => Boolean(tokenId))
-  }, [selectedMarket])
-
-  const isSelectedMarketResolved = Boolean(selectedMarket?.is_resolved || selectedMarket?.condition?.resolved)
-
-  return {
-    selectedButton,
-    selectedMarket,
-    selectedOutcome,
-    selectedLinePickerMarketType,
-    nextOutcome,
-    nextButton,
-    tradeSelectionLabel,
-    switchTooltip,
-    selectedMarketTokenIds,
-    isSelectedMarketResolved,
-  }
-}
-
-function useSportsLinePicker({
-  card,
-  allowedConditionIds,
-  selectedLinePickerMarketType,
-  selectedButton,
-  onSelectButton,
-}: {
-  card: SportsGamesCard
-  allowedConditionIds: Set<string> | null
-  selectedLinePickerMarketType: LinePickerMarketType | null
-  selectedButton: SportsGamesButton | null
-  onSelectButton: (buttonKey: string, options?: { panelMode?: 'full' | 'partial' | 'preserve' }) => void
-}) {
-  const linePickerScrollerRef = useRef<HTMLDivElement | null>(null)
-  const linePickerButtonsRef = useRef<Record<string, HTMLButtonElement | null>>({})
-  const linePickerScrollSettleTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const linePickerSuppressScrollSyncUntilRef = useRef(0)
-
-  const linePickerOptions = useMemo(
-    () => {
-      if (!selectedLinePickerMarketType) {
-        return []
-      }
-
-      const options = buildLinePickerOptions(card, selectedLinePickerMarketType)
-      if (!allowedConditionIds) {
-        return options
-      }
-
-      return options.filter(option => allowedConditionIds.has(option.conditionId))
-    },
-    [allowedConditionIds, card, selectedLinePickerMarketType],
-  )
-
-  const activeLineOptionIndex = useMemo(() => {
-    if (!selectedButton || linePickerOptions.length === 0) {
-      return -1
-    }
-
-    return linePickerOptions.findIndex(option => option.conditionId === selectedButton.conditionId)
-  }, [linePickerOptions, selectedButton])
-
-  const hasLinePicker = selectedLinePickerMarketType !== null && linePickerOptions.length > 1
-
-  const suppressLinePickerScrollSync = useCallback((durationMs = 220) => {
-    linePickerSuppressScrollSyncUntilRef.current = Date.now() + durationMs
-  }, [])
-
-  const pickLineOption = useCallback((optionIndex: number) => {
-    if (!selectedButton) {
-      return
-    }
-
-    const option = linePickerOptions[optionIndex]
-    if (!option) {
-      return
-    }
-
-    const preferredButton = resolvePreferredLinePickerButton(option.buttons, selectedButton)
-    if (!preferredButton) {
-      return
-    }
-
-    suppressLinePickerScrollSync()
-    onSelectButton(preferredButton.key, { panelMode: 'preserve' })
-  }, [linePickerOptions, onSelectButton, selectedButton, suppressLinePickerScrollSync])
-
-  const handlePickPreviousLine = useCallback(() => {
-    if (activeLineOptionIndex <= 0) {
-      return
-    }
-    pickLineOption(activeLineOptionIndex - 1)
-  }, [activeLineOptionIndex, pickLineOption])
-
-  const handlePickNextLine = useCallback(() => {
-    if (activeLineOptionIndex < 0 || activeLineOptionIndex >= linePickerOptions.length - 1) {
-      return
-    }
-    pickLineOption(activeLineOptionIndex + 1)
-  }, [activeLineOptionIndex, linePickerOptions.length, pickLineOption])
-
-  const resolveCenteredLineOptionIndex = useCallback(() => {
-    const scroller = linePickerScrollerRef.current
-    if (!scroller || linePickerOptions.length === 0) {
-      return -1
-    }
-
-    const scrollerCenter = scroller.scrollLeft + scroller.clientWidth / 2
-    let closestIndex = -1
-    let smallestDistance = Number.POSITIVE_INFINITY
-
-    linePickerOptions.forEach((option, index) => {
-      const button = linePickerButtonsRef.current[option.conditionId]
-      if (!button) {
-        return
-      }
-
-      const buttonCenter = button.offsetLeft + button.offsetWidth / 2
-      const distance = Math.abs(buttonCenter - scrollerCenter)
-      if (distance < smallestDistance) {
-        smallestDistance = distance
-        closestIndex = index
-      }
-    })
-
-    return closestIndex
-  }, [linePickerOptions])
-
-  const alignActiveLineOption = useCallback((behavior: ScrollBehavior = 'smooth') => {
-    if (activeLineOptionIndex < 0) {
-      return
-    }
-
-    const scroller = linePickerScrollerRef.current
-    if (!scroller) {
-      return
-    }
-
-    const activeOption = linePickerOptions[activeLineOptionIndex]
-    if (!activeOption) {
-      return
-    }
-
-    const activeButton = linePickerButtonsRef.current[activeOption.conditionId]
-    if (!activeButton) {
-      return
-    }
-
-    suppressLinePickerScrollSync()
-    const targetLeft = activeButton.offsetLeft - ((scroller.clientWidth - activeButton.offsetWidth) / 2)
-    scroller.scrollTo({
-      left: Math.max(0, targetLeft),
-      behavior,
-    })
-  }, [activeLineOptionIndex, linePickerOptions, suppressLinePickerScrollSync])
-
-  useEffect(function alignActiveLineOptionOnIndexChange() {
-    if (activeLineOptionIndex < 0) {
-      return undefined
-    }
-    alignActiveLineOption('auto')
-    return undefined
-  }, [activeLineOptionIndex, alignActiveLineOption])
-
-  useEffect(function alignActiveLineOptionOnPickerMount() {
-    if (!hasLinePicker) {
-      return undefined
-    }
-
-    const frame = window.requestAnimationFrame(() => {
-      alignActiveLineOption('auto')
-    })
-
-    return function cancelAlignActiveLineOptionFrame() {
-      window.cancelAnimationFrame(frame)
-    }
-  }, [alignActiveLineOption, hasLinePicker])
-
-  useEffect(function syncLinePickerScrollToCenteredOption() {
-    const scrollerElement = linePickerScrollerRef.current
-    if (!hasLinePicker || !scrollerElement) {
-      return undefined
-    }
-
-    function syncCenteredLineOption() {
-      const centeredIndex = resolveCenteredLineOptionIndex()
-      if (centeredIndex < 0 || centeredIndex === activeLineOptionIndex) {
-        return
-      }
-
-      pickLineOption(centeredIndex)
-    }
-
-    function handleScroll() {
-      if (Date.now() < linePickerSuppressScrollSyncUntilRef.current) {
-        return
-      }
-
-      if (linePickerScrollSettleTimeoutRef.current) {
-        clearTimeout(linePickerScrollSettleTimeoutRef.current)
-      }
-
-      linePickerScrollSettleTimeoutRef.current = setTimeout(() => {
-        linePickerScrollSettleTimeoutRef.current = null
-        syncCenteredLineOption()
-      }, 90)
-    }
-
-    scrollerElement.addEventListener('scroll', handleScroll, { passive: true })
-
-    return function detachLinePickerScrollSync() {
-      scrollerElement.removeEventListener('scroll', handleScroll)
-      if (linePickerScrollSettleTimeoutRef.current) {
-        clearTimeout(linePickerScrollSettleTimeoutRef.current)
-        linePickerScrollSettleTimeoutRef.current = null
-      }
-    }
-  }, [activeLineOptionIndex, hasLinePicker, pickLineOption, resolveCenteredLineOptionIndex])
-
-  return {
-    linePickerScrollerRef,
-    linePickerButtonsRef,
-    linePickerOptions,
-    activeLineOptionIndex,
-    hasLinePicker,
-    pickLineOption,
-    handlePickPreviousLine,
-    handlePickNextLine,
-  }
-}
-
-function useSportsDetailsTabs({
-  activeDetailsTab,
-  showBottomContent,
-  showAboutTab,
-  aboutEvent,
-  isSelectedMarketResolved,
-  onChangeTab,
-}: {
-  activeDetailsTab: DetailsTab
-  showBottomContent: boolean
-  showAboutTab: boolean
-  aboutEvent: SportsGamesCard['event'] | null
-  isSelectedMarketResolved: boolean
-  onChangeTab: (tab: DetailsTab) => void
-}) {
-  const detailTabs = useMemo<Array<{ id: DetailsTab, label: string }>>(() => {
-    const tabs: Array<{ id: DetailsTab, label: string }> = []
-
-    if (!isSelectedMarketResolved) {
-      tabs.push({ id: 'orderBook', label: 'Order Book' })
-    }
-
-    tabs.push({ id: 'graph', label: 'Graph' })
-
-    if (showAboutTab && aboutEvent) {
-      tabs.push({ id: 'about', label: 'About' })
-    }
-
-    return tabs
-  }, [aboutEvent, isSelectedMarketResolved, showAboutTab])
-
-  const resolvedActiveDetailsTab = useMemo<DetailsTab>(() => {
-    if (detailTabs.some(tab => tab.id === activeDetailsTab)) {
-      return activeDetailsTab
-    }
-
-    return detailTabs[0]?.id ?? 'orderBook'
-  }, [activeDetailsTab, detailTabs])
-
-  useEffect(function syncResolvedDetailsTabUpstream() {
-    if (!showBottomContent) {
-      return undefined
-    }
-
-    if (resolvedActiveDetailsTab !== activeDetailsTab) {
-      onChangeTab(resolvedActiveDetailsTab)
-    }
-
-    return undefined
-  }, [activeDetailsTab, onChangeTab, resolvedActiveDetailsTab, showBottomContent])
-
-  return { detailTabs, resolvedActiveDetailsTab }
-}
-
-function useSportsOwnerAddress() {
-  const user = useUser()
-  const ownerAddress = useMemo(() => {
-    if (user?.proxy_wallet_address && user.proxy_wallet_status === 'deployed') {
-      return user.proxy_wallet_address
-    }
-    return null
-  }, [user?.proxy_wallet_address, user?.proxy_wallet_status])
-
-  return ownerAddress
-}
-
-function useSportsPositionOddsFormatters(oddsFormat: OddsFormat) {
-  const formatPositionOddsLabel = useCallback((cents: number | null) => {
-    if (oddsFormat === 'price') {
-      return formatCompactCentsLabel(cents)
-    }
-    return formatOddsFromCents(cents, oddsFormat)
-  }, [oddsFormat])
-
-  const formatAverageCellLabel = useCallback((cents: number | null) => {
-    if (oddsFormat === 'price') {
-      return formatCentsLabel(cents, { fallback: '—' })
-    }
-    return formatOddsFromCents(cents, oddsFormat)
-  }, [oddsFormat])
-
-  return { formatPositionOddsLabel, formatAverageCellLabel }
-}
-
-function useSportsCashOutHandlers({
-  card,
-  isMobile,
-  setCashOutPayload,
-  orderStore,
-}: {
-  card: SportsGamesCard
-  isMobile: boolean
-  setCashOutPayload: (payload: SportsCashOutModalPayload | null) => void
-  orderStore: ReturnType<typeof useSportsGameDetailsPanelOrderStore>
-}) {
-  const { setOrderType, setOrderSide, setOrderMarket, setOrderOutcome, setOrderAmount, setIsMobileOrderPanelOpen } = orderStore
-
-  const handleCashOutTag = useCallback(async (
-    tag: SportsPositionTag,
-    event?: ReactMouseEventType<HTMLElement>,
-  ) => {
-    event?.stopPropagation()
-
-    const tokenId = tag.outcome.token_id ? String(tag.outcome.token_id) : null
-    if (!tokenId) {
-      return
-    }
-
-    let summary = await fetchOrderBookSummaries([tokenId])
-      .then(result => result[tokenId])
-      .catch(() => null)
-
-    if (!summary) {
-      summary = null
-    }
-
-    const bids = normalizeBookLevels(summary?.bids, 'bid')
-    const asks = normalizeBookLevels(summary?.asks, 'ask')
-    const fill = calculateMarketFill(ORDER_SIDE.SELL, tag.shares, bids, asks)
-
-    setOrderType(ORDER_TYPE.MARKET)
-    setOrderSide(ORDER_SIDE.SELL)
-    setOrderMarket(tag.market)
-    setOrderOutcome(tag.outcome)
-    setOrderAmount(formatAmountInputValue(tag.shares, { roundingMode: 'floor' }))
-
-    if (isMobile) {
-      setIsMobileOrderPanelOpen(true)
-    }
-
-    setCashOutPayload({
-      outcomeLabel: tag.summaryLabel,
-      outcomeShortLabel: card.event.title || tag.market.short_title || tag.market.title,
-      outcomeIconUrl: tag.market.icon_url,
-      shares: tag.shares,
-      filledShares: fill.filledShares,
-      avgPriceCents: fill.avgPriceCents,
-      receiveAmount: fill.totalCost > 0 ? fill.totalCost : null,
-      sellBids: bids,
-    })
-  }, [
-    card.event.title,
-    isMobile,
-    setCashOutPayload,
-    setIsMobileOrderPanelOpen,
-    setOrderAmount,
-    setOrderMarket,
-    setOrderOutcome,
-    setOrderSide,
-    setOrderType,
-  ])
-
-  const handleCashOutModalChange = useCallback((open: boolean) => {
-    if (!open) {
-      setCashOutPayload(null)
-    }
-  }, [setCashOutPayload])
-
-  const handleCashOutSubmit = useCallback((sharesToSell: number) => {
-    if (!(sharesToSell > 0)) {
-      return
-    }
-    setOrderAmount(formatAmountInputValue(sharesToSell, { roundingMode: 'floor' }))
-    setCashOutPayload(null)
-    const form = document.getElementById('event-order-form') as HTMLFormElement | null
-    form?.requestSubmit()
-  }, [setCashOutPayload, setOrderAmount])
-
-  return { handleCashOutTag, handleCashOutModalChange, handleCashOutSubmit }
-}
-
-function useSportsDetailsPanelInteractions({
-  selectedMarket,
-  nextOutcome,
-  nextButton,
-  onSelectButton,
-  setOrderMarket,
-  setOrderOutcome,
-  isNegRiskEnabled,
-  moneylineConditionIds,
-  setConvertTagKey,
-}: {
-  selectedMarket: Market | null
-  nextOutcome: Outcome | null
-  nextButton: SportsGamesButton | null
-  onSelectButton: SportsGameDetailsPanelProps['onSelectButton']
-  setOrderMarket: (market: Market) => void
-  setOrderOutcome: (outcome: Outcome) => void
-  isNegRiskEnabled: boolean
-  moneylineConditionIds: Set<string>
-  setConvertTagKey: (key: string | null) => void
-}) {
-  const handleToggleOutcome = useCallback(() => {
-    if (!selectedMarket || !nextOutcome) {
-      return
-    }
-
-    setOrderMarket(selectedMarket)
-    setOrderOutcome(nextOutcome)
-    if (nextButton) {
-      onSelectButton(nextButton.key, { panelMode: 'preserve' })
-    }
-  }, [nextButton, nextOutcome, onSelectButton, selectedMarket, setOrderMarket, setOrderOutcome])
-
-  const handleOpenConvert = useCallback((
-    tag: SportsPositionTag,
-    event?: ReactMouseEventType<HTMLElement>,
-  ) => {
-    event?.stopPropagation()
-    if (
-      !isNegRiskEnabled
-      || !moneylineConditionIds.has(tag.conditionId)
-      || tag.outcomeIndex !== OUTCOME_INDEX.NO
-      || tag.outcome.outcome_index !== OUTCOME_INDEX.NO
-      || tag.shares <= 0
-    ) {
-      return
-    }
-    setConvertTagKey(tag.key)
-  }, [isNegRiskEnabled, moneylineConditionIds, setConvertTagKey])
-
-  return { handleToggleOutcome, handleOpenConvert }
-}
-
-export function SportsGameDetailsPanel({
-  card,
-  activeDetailsTab,
-  selectedButtonKey,
-  showBottomContent,
-  defaultGraphTimeRange = '1W',
-  allowedConditionIds = null,
-  positionsTitle,
-  showAboutTab = false,
-  aboutEvent = null,
-  rulesEvent = null,
-  showRedeemInPositions = false,
-  onOpenRedeemForCondition = null,
-  oddsFormat = 'price',
-  marketContextEnabled = true,
-  onChangeTab,
-  onSelectButton,
-}: SportsGameDetailsPanelProps) {
-  const t = useExtracted()
-  const isMobile = useIsMobile()
-  const linePickerSpacerWidth = 'calc(50% - 28px)'
-
-  const orderStore = useSportsGameDetailsPanelOrderStore()
-  const {
-    orderMarketConditionId,
-    orderOutcomeIndex,
-    setOrderMarket,
-    setOrderOutcome,
-    setOrderAmount,
-  } = orderStore
-
-  const {
-    cashOutPayload,
-    setCashOutPayload,
-    isPositionsExpanded,
-    setIsPositionsExpanded,
-    convertTagKey,
-    setConvertTagKey,
-  } = useSportsGameDetailsPanelLocalState()
-
-  const ownerAddress = useSportsOwnerAddress()
-
-  const {
-    cardMarketByConditionId,
-    cardButtonsByConditionAndOutcome,
-    cardFirstButtonByCondition,
-    moneylineConditionIds,
-    isNegRiskEnabled,
-  } = useSportsCardDerivations(card)
-
-  const userPositions = useSportsCardUserPositionsQuery({
-    ownerAddress,
-    cardId: card.id,
-    showBottomContent,
-  })
-
-  const { positionTags, visiblePositionTags, hiddenPositionTagsCount } = useSportsPositionTags({
-    ownerAddress,
-    userPositions,
-    allowedConditionIds,
-    card,
-    cardMarketByConditionId,
-    cardButtonsByConditionAndOutcome,
-    cardFirstButtonByCondition,
-  })
-
-  const { convertDialogTag, convertDialogOptions, convertDialogOutcomes } = useSportsConvertDialog({
-    convertTagKey,
-    positionTags,
-    card,
-    allowedConditionIds,
-  })
-
-  const {
-    selectedButton,
-    selectedMarket,
-    selectedOutcome,
-    selectedLinePickerMarketType,
-    nextOutcome,
-    nextButton,
-    tradeSelectionLabel,
-    switchTooltip,
-    selectedMarketTokenIds,
-    isSelectedMarketResolved,
-  } = useSportsSelectedMarketDerivations({
-    card,
-    selectedButtonKey,
-    orderMarketConditionId,
-    orderOutcomeIndex,
-  })
-
-  const {
-    linePickerScrollerRef,
-    linePickerButtonsRef,
-    linePickerOptions,
-    activeLineOptionIndex,
-    hasLinePicker,
-    pickLineOption,
-    handlePickPreviousLine,
-    handlePickNextLine,
-  } = useSportsLinePicker({
-    card,
-    allowedConditionIds,
-    selectedLinePickerMarketType,
-    selectedButton,
-    onSelectButton,
-  })
-
-  const { detailTabs, resolvedActiveDetailsTab } = useSportsDetailsTabs({
-    activeDetailsTab,
-    showBottomContent,
-    showAboutTab,
-    aboutEvent,
-    isSelectedMarketResolved,
-    onChangeTab,
-  })
-
-  const {
-    data: orderBookSummaries,
-    isLoading: isOrderBookLoading,
-    isRefetching: isOrderBookRefetching,
-    refetch: refetchOrderBook,
-  } = useOrderBookSummaries(selectedMarketTokenIds, {
-    enabled: showBottomContent && activeDetailsTab === 'orderBook' && selectedMarketTokenIds.length > 0,
-  })
-
-  const { formatPositionOddsLabel, formatAverageCellLabel } = useSportsPositionOddsFormatters(oddsFormat)
-
-  const { handleCashOutTag, handleCashOutModalChange, handleCashOutSubmit } = useSportsCashOutHandlers({
-    card,
-    isMobile,
-    setCashOutPayload,
-    orderStore,
-  })
-
-  const { handleToggleOutcome, handleOpenConvert } = useSportsDetailsPanelInteractions({
-    selectedMarket,
-    nextOutcome,
-    nextButton,
-    onSelectButton,
-    setOrderMarket,
-    setOrderOutcome,
-    isNegRiskEnabled,
-    moneylineConditionIds,
-    setConvertTagKey,
-  })
-
-  const isStandalonePositionsCard = Boolean(positionsTitle)
-  const shouldShowPortfolio = visiblePositionTags.length > 0
-  const showPositionTagSummary = !isStandalonePositionsCard
-
-  if (!showBottomContent && !hasLinePicker && !shouldShowPortfolio) {
-    return null
-  }
-
-  return (
-    <>
-      <div
-        className={cn(
-          'overflow-x-visible overflow-y-hidden transition-[max-height,opacity,margin] duration-200',
-          hasLinePicker
-            ? (showBottomContent ? '-mt-3 mb-3 max-h-32 opacity-100' : '-mt-3 mb-0 max-h-32 opacity-100')
-            : 'mb-0 max-h-0 opacity-0',
-        )}
-      >
-        {hasLinePicker && (
-          <div className={cn(
-            '-mx-2.5 bg-card px-2.5',
-            showBottomContent ? 'pb-0' : 'pb-2',
-          )}
-          >
-            {!showBottomContent && <div className="-mx-2.5 border-t" />}
-
-            <div className="pt-2">
-              <div className="mt-0.5 flex items-center gap-2">
-                <button
-                  type="button"
-                  onClick={handlePickPreviousLine}
-                  disabled={activeLineOptionIndex <= 0}
-                  className={cn(
-                    `
-                      inline-flex size-7 items-center justify-center rounded-sm text-muted-foreground transition-colors
-                      focus-visible:ring-1 focus-visible:ring-ring focus-visible:outline-none
-                    `,
-                    activeLineOptionIndex > 0
-                      ? 'cursor-pointer hover:bg-muted/70 hover:text-foreground'
-                      : 'cursor-not-allowed opacity-40',
-                  )}
-                  aria-label="Previous line"
-                >
-                  <ChevronLeftIcon className="size-4.5" />
-                </button>
-
-                <div
-                  className="relative min-w-0 flex-1"
-                >
-                  <span
-                    aria-hidden
-                    className="
-                      pointer-events-none absolute -top-2 left-1/2 h-2 w-3 -translate-x-1/2 bg-primary
-                      [clip-path:polygon(50%_100%,0_0,100%_0)]
-                    "
-                  />
-
-                  <div
-                    ref={linePickerScrollerRef}
-                    className={`
-                      flex min-w-0 snap-x snap-mandatory items-center gap-2 overflow-x-auto scroll-smooth
-                      [scrollbar-width:none]
-                      [&::-webkit-scrollbar]:hidden
-                    `}
-                  >
-                    <span aria-hidden className="shrink-0" style={{ width: linePickerSpacerWidth }} />
-                    {linePickerOptions.map((option, index) => (
-                      <button
-                        key={`${card.id}-${option.conditionId}`}
-                        type="button"
-                        onClick={() => pickLineOption(index)}
-                        ref={(node) => {
-                          linePickerButtonsRef.current[option.conditionId] = node
-                        }}
-                        className={cn(
-                          `
-                            w-10 shrink-0 snap-center text-center text-sm font-medium text-muted-foreground
-                            transition-colors
-                          `,
-                          index === activeLineOptionIndex
-                            ? 'text-base font-semibold text-foreground'
-                            : 'hover:text-foreground/80',
-                        )}
-                      >
-                        {option.label}
-                      </button>
-                    ))}
-                    <span aria-hidden className="shrink-0" style={{ width: linePickerSpacerWidth }} />
-                  </div>
-                </div>
-
-                <button
-                  type="button"
-                  onClick={handlePickNextLine}
-                  disabled={activeLineOptionIndex < 0 || activeLineOptionIndex >= linePickerOptions.length - 1}
-                  className={cn(
-                    `
-                      inline-flex size-7 items-center justify-center rounded-sm text-muted-foreground transition-colors
-                      focus-visible:ring-1 focus-visible:ring-ring focus-visible:outline-none
-                    `,
-                    activeLineOptionIndex >= 0 && activeLineOptionIndex < linePickerOptions.length - 1
-                      ? 'cursor-pointer hover:bg-muted/70 hover:text-foreground'
-                      : 'cursor-not-allowed opacity-40',
-                  )}
-                  aria-label="Next line"
-                >
-                  <ChevronRightIcon className="size-4.5" />
-                </button>
-              </div>
-            </div>
-
-            {showBottomContent && (
-              <div className="-mx-2.5 mt-2 border-t" />
-            )}
-          </div>
-        )}
-      </div>
-
-      {showBottomContent && (
-        <>
-          <div className="-mx-2.5 mb-3 border-b bg-card">
-            <div className="flex w-full items-center gap-2 px-2.5">
-              <div className="flex w-0 flex-1 items-center gap-4 overflow-x-auto">
-                {detailTabs.map(tab => (
-                  <button
-                    key={`${card.id}-${tab.id}`}
-                    type="button"
-                    onClick={() => onChangeTab(tab.id)}
-                    className={cn(
-                      `
-                        border-b-2 border-transparent pt-1 pb-2 text-sm font-semibold whitespace-nowrap
-                        transition-colors
-                      `,
-                      resolvedActiveDetailsTab === tab.id
-                        ? 'border-primary text-foreground'
-                        : 'text-muted-foreground hover:text-foreground',
-                    )}
-                  >
-                    {tab.label}
-                  </button>
-                ))}
-              </div>
-
-              {selectedMarketTokenIds.length > 0 && resolvedActiveDetailsTab !== 'about' && (
-                <button
-                  type="button"
-                  className={cn(
-                    `
-                      -mt-1 ml-auto inline-flex size-7 items-center justify-center rounded-sm text-muted-foreground
-                      transition-colors
-                    `,
-                    'hover:bg-muted/70 hover:text-foreground',
-                    'focus-visible:ring-1 focus-visible:ring-ring focus-visible:outline-none',
-                  )}
-                  aria-label="Refresh order book"
-                  title="Refresh order book"
-                  onClick={() => { void refetchOrderBook() }}
-                  disabled={isOrderBookLoading || isOrderBookRefetching}
-                >
-                  <RefreshCwIcon
-                    className={cn(
-                      'size-3',
-                      { 'animate-spin': isOrderBookLoading || isOrderBookRefetching },
-                    )}
-                  />
-                </button>
-              )}
-            </div>
-          </div>
-
-          {resolvedActiveDetailsTab === 'orderBook' && (
-            (selectedMarket && selectedOutcome)
-              ? (
-                  <div className={cn('-mx-2.5', visiblePositionTags.length === 0 && '-mb-2.5')}>
-                    <EventOrderBook
-                      market={selectedMarket}
-                      outcome={selectedOutcome}
-                      summaries={orderBookSummaries}
-                      isLoadingSummaries={isOrderBookLoading && !orderBookSummaries}
-                      eventSlug={card.slug}
-                      surfaceVariant="sportsCard"
-                      oddsFormat={oddsFormat}
-                      tradeLabel={`TRADE ${tradeSelectionLabel}`}
-                      onToggleOutcome={nextOutcome ? handleToggleOutcome : undefined}
-                      toggleOutcomeTooltip={switchTooltip ?? undefined}
-                      openMobileOrderPanelOnLevelSelect={isMobile}
-                    />
-                  </div>
-                )
-              : (
-                  <div className="rounded-lg border bg-card px-3 py-6 text-sm text-muted-foreground">
-                    Order book is unavailable for this game.
-                  </div>
-                )
-          )}
-
-          {resolvedActiveDetailsTab === 'graph' && (
-            <SportsGameGraph
-              key={`${card.id}:${selectedButton?.marketType ?? 'moneyline'}:${selectedButton?.conditionId ?? 'none'}:${defaultGraphTimeRange}`}
-              card={card}
-              selectedMarketType={selectedButton?.marketType ?? 'moneyline'}
-              selectedConditionId={selectedButton?.conditionId ?? null}
-              defaultTimeRange={defaultGraphTimeRange}
-              variant="sportsCardLegend"
-            />
-          )}
-
-          {resolvedActiveDetailsTab === 'about' && aboutEvent && (
-            <SportsEventAboutPanel
-              event={aboutEvent}
-              rulesEvent={rulesEvent}
-              market={selectedMarket}
-              marketContextEnabled={marketContextEnabled}
-            />
-          )}
-        </>
-      )}
-
-      {shouldShowPortfolio && (
-        <div className={cn(
-          '-mx-2.5 bg-card',
-          isStandalonePositionsCard && 'overflow-hidden rounded-[inherit]',
-        )}
-        >
-          <div className={cn(!isStandalonePositionsCard && 'border-t')}>
-            <div
-              role="button"
-              tabIndex={0}
-              data-sports-card-control="true"
-              onClick={(event) => {
-                event.stopPropagation()
-                setIsPositionsExpanded(current => !current)
-              }}
-              onKeyDown={(event) => {
-                if (event.key !== 'Enter' && event.key !== ' ') {
-                  return
-                }
-                event.preventDefault()
-                event.stopPropagation()
-                setIsPositionsExpanded(current => !current)
-              }}
-              className={cn(
-                'flex w-full items-center bg-card text-muted-foreground transition-colors hover:bg-secondary',
-                isStandalonePositionsCard
-                  ? 'min-h-16 gap-3 px-4 py-3 text-sm'
-                  : 'min-h-11 gap-2 px-2.5 py-2 text-xs sm:px-2.5',
-              )}
-            >
-              <div
-                className={cn(
-                  'flex shrink-0 items-center text-foreground',
-                  isStandalonePositionsCard ? 'text-sm font-semibold' : 'text-sm font-semibold',
-                )}
-              >
-                <span>{positionsTitle ?? t('Positions')}</span>
-              </div>
-
-              {showPositionTagSummary && (
-                <>
-                  <div className="ml-auto flex min-w-0 flex-1 items-center justify-end gap-1 overflow-hidden">
-                    {visiblePositionTags.map((tag) => {
-                      const tagAccent = tag.button
-                        ? resolveTradeHeaderBadgeAccent(tag.button)
-                        : (tag.outcomeIndex === OUTCOME_INDEX.NO
-                            ? { className: 'bg-no/10 text-no', style: undefined }
-                            : { className: 'bg-yes/10 text-yes', style: undefined })
-
-                      return (
-                        <span
-                          key={tag.key}
-                          className={cn(
-                            `
-                              group/position inline-flex max-w-44 min-w-0 items-center rounded-sm px-2.5 py-1 text-xs
-                              font-semibold
-                            `,
-                            tagAccent.className,
-                          )}
-                          style={tagAccent.style}
-                        >
-                          <span className="truncate whitespace-nowrap">
-                            {`${tag.summaryLabel} | ${formatSharesLabel(tag.shares)} @ ${formatPositionOddsLabel(tag.avgPriceCents)}`}
-                          </span>
-                          <button
-                            type="button"
-                            data-sports-card-control="true"
-                            className={cn(
-                              'ml-1 inline-flex w-0 items-center justify-center overflow-hidden opacity-0',
-                              'transition-all duration-150 group-hover/position:w-3 group-hover/position:opacity-100',
-                              'pointer-events-none group-hover/position:pointer-events-auto',
-                            )}
-                            aria-label={`Cash out ${tag.summaryLabel}`}
-                            onClick={event => void handleCashOutTag(tag, event)}
-                          >
-                            <XIcon className="size-3" />
-                          </button>
-                        </span>
-                      )
-                    })}
-                  </div>
-
-                  {hiddenPositionTagsCount > 0 && (
-                    <span className="shrink-0 text-xs font-semibold text-muted-foreground">
-                      {`+${hiddenPositionTagsCount} more`}
-                    </span>
-                  )}
-                </>
-              )}
-
-              <ChevronDownIcon
-                className={cn(
-                  'shrink-0 transition-transform',
-                  !showPositionTagSummary && 'ml-auto',
-                  isStandalonePositionsCard ? 'size-4' : 'size-3.5',
-                  isPositionsExpanded ? 'rotate-180' : 'rotate-0',
-                )}
-              />
-            </div>
-
-            {isPositionsExpanded && (
-              <div className="border-t bg-card px-2.5 py-2 sm:px-2.5" data-sports-card-control="true">
-                <div className="w-full overflow-x-auto" onClick={event => event.stopPropagation()}>
-                  <table className="w-full border-collapse text-xs">
-                    <thead>
-                      <tr className="text-2xs font-semibold tracking-wide text-muted-foreground uppercase">
-                        <th className="py-2 text-left">Type</th>
-                        <th className="p-2 text-left">Outcome</th>
-                        <th className="p-2 text-right">Avg</th>
-                        <th className="p-2 text-right">Cost</th>
-                        <th className="p-2 text-right">To Win</th>
-                        <th className="p-2 text-right">Current</th>
-                        <th className="py-2 text-right" />
-                      </tr>
-                      <tr>
-                        <th colSpan={7} className="p-0">
-                          <div className="border-t" />
-                        </th>
-                      </tr>
-                    </thead>
-                    <tbody className="divide-y divide-border/60">
-                      {positionTags.map((tag) => {
-                        const tagAccent = tag.button
-                          ? resolveTradeHeaderBadgeAccent(tag.button)
-                          : (tag.outcomeIndex === OUTCOME_INDEX.NO
-                              ? { className: 'bg-no/10 text-no', style: undefined }
-                              : { className: 'bg-yes/10 text-yes', style: undefined })
-                        const costLabel = typeof tag.totalCost === 'number'
-                          ? formatCurrency(tag.totalCost, { minimumFractionDigits: 2, maximumFractionDigits: 2 })
-                          : '—'
-                        const toWinLabel = formatCurrency(tag.shares, { minimumFractionDigits: 2, maximumFractionDigits: 2 })
-                        const currentLabel = formatCurrency(tag.currentValue, { minimumFractionDigits: 2, maximumFractionDigits: 2 })
-                        const pnlValue = typeof tag.totalCost === 'number'
-                          ? tag.currentValue - tag.totalCost + tag.realizedPnl
-                          : null
-                        const pnlLabel = pnlValue == null
-                          ? '—'
-                          : `${pnlValue >= 0 ? '+' : '-'}${formatCurrency(Math.abs(pnlValue), { minimumFractionDigits: 2, maximumFractionDigits: 2 })}`
-                        const pnlClass = pnlValue == null
-                          ? (tag.currentValue >= 0 ? 'text-yes' : 'text-no')
-                          : pnlValue >= 0
-                            ? 'text-yes'
-                            : 'text-no'
-                        const canConvert = isNegRiskEnabled
-                          && moneylineConditionIds.has(tag.conditionId)
-                          && tag.outcomeIndex === OUTCOME_INDEX.NO
-                          && tag.outcome.outcome_index === OUTCOME_INDEX.NO
-                          && tag.shares > 0
-                        const canRedeem = showRedeemInPositions
-                          && Boolean(tag.market.is_resolved || tag.market.condition?.resolved)
-
-                        return (
-                          <tr key={tag.key} className="text-xs text-foreground">
-                            <td className="py-2 font-medium">{tag.marketTypeLabel}</td>
-                            <td className="p-2">
-                              <span
-                                className={cn(
-                                  'inline-flex min-w-0 items-center rounded-sm px-2.5 py-1 text-xs font-semibold',
-                                  tagAccent.className,
-                                )}
-                                style={tagAccent.style}
-                              >
-                                {`${tag.summaryLabel} | ${formatSharesLabel(tag.shares)}`}
-                              </span>
-                            </td>
-                            <td className="p-2 text-right font-medium">
-                              {formatAverageCellLabel(tag.avgPriceCents)}
-                            </td>
-                            <td className="p-2 text-right font-medium">{costLabel}</td>
-                            <td className="p-2 text-right font-medium">{toWinLabel}</td>
-                            <td className={cn('p-2 text-right font-medium', pnlClass)}>
-                              {currentLabel}
-                              {' '}
-                              (
-                              {pnlLabel}
-                              )
-                            </td>
-                            <td className="py-2 text-right">
-                              <div className="flex items-center justify-end gap-1.5">
-                                {canConvert && (
-                                  <button
-                                    type="button"
-                                    data-sports-card-control="true"
-                                    className={`
-                                      inline-flex h-7 items-center justify-center rounded-sm bg-secondary/70 px-2
-                                      text-xs font-semibold text-foreground transition-colors
-                                      hover:bg-secondary
-                                    `}
-                                    onClick={event => handleOpenConvert(tag, event)}
-                                  >
-                                    Convert
-                                  </button>
-                                )}
-                                {canRedeem
-                                  ? (
-                                      <button
-                                        type="button"
-                                        data-sports-card-control="true"
-                                        className={`
-                                          inline-flex h-7 items-center justify-center rounded-sm border border-border/70
-                                          bg-background px-2 text-xs font-semibold text-foreground transition-colors
-                                          hover:bg-secondary/35
-                                        `}
-                                        onClick={(event) => {
-                                          event.stopPropagation()
-                                          onOpenRedeemForCondition?.(tag.conditionId)
-                                        }}
-                                      >
-                                        Redeem
-                                      </button>
-                                    )
-                                  : (
-                                      <button
-                                        type="button"
-                                        data-sports-card-control="true"
-                                        className={`
-                                          inline-flex h-7 items-center justify-center rounded-sm border border-border/70
-                                          bg-background/40 px-2 text-xs font-semibold text-foreground transition-colors
-                                          hover:bg-secondary/40
-                                        `}
-                                        onClick={event => void handleCashOutTag(tag, event)}
-                                      >
-                                        Sell
-                                      </button>
-                                    )}
-                              </div>
-                            </td>
-                          </tr>
-                        )
-                      })}
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-
-      {convertDialogTag && (
-        <EventConvertPositionsDialog
-          open={Boolean(convertDialogTag)}
-          options={convertDialogOptions}
-          outcomes={convertDialogOutcomes}
-          negRiskMarketId={card.event.neg_risk_market_id ?? undefined}
-          isNegRiskAugmented={Boolean(card.event.neg_risk_augmented)}
-          onOpenChange={(open) => {
-            if (!open) {
-              setConvertTagKey(null)
-            }
-          }}
-        />
-      )}
-
-      {cashOutPayload && (
-        <SellPositionModal
-          open={Boolean(cashOutPayload)}
-          onOpenChange={handleCashOutModalChange}
-          outcomeLabel={cashOutPayload.outcomeLabel}
-          outcomeShortLabel={cashOutPayload.outcomeShortLabel}
-          outcomeIconUrl={cashOutPayload.outcomeIconUrl}
-          fallbackIconUrl={card.event.icon_url}
-          shares={cashOutPayload.shares}
-          filledShares={cashOutPayload.filledShares}
-          avgPriceCents={cashOutPayload.avgPriceCents}
-          receiveAmount={cashOutPayload.receiveAmount}
-          sellBids={cashOutPayload.sellBids}
-          onSharesChange={sharesToSell =>
-            setOrderAmount(formatAmountInputValue(sharesToSell, { roundingMode: 'floor' }))}
-          onCashOut={handleCashOutSubmit}
-          onEditOrder={(sharesToSell) => {
-            setOrderAmount(formatAmountInputValue(sharesToSell, { roundingMode: 'floor' }))
-            setCashOutPayload(null)
-          }}
-        />
-      )}
-    </>
-  )
-}
-
-function useSportsGamesCenterShellState() {
-  const [openCardId, setOpenCardId] = useState<string | null>(null)
-  const [isDetailsContentVisible, setIsDetailsContentVisible] = useState(true)
-  const [activeDetailsTab, setActiveDetailsTab] = useState<DetailsTab>('orderBook')
-  const [selectedConditionByCardId, setSelectedConditionByCardId] = useState<Record<string, string>>({})
-  const [tradeSelection, setTradeSelection] = useState<SportsTradeSelection>({ cardId: null, buttonKey: null })
-  const [isSearchOpen, setIsSearchOpen] = useState(false)
-  const [searchQuery, setSearchQuery] = useState('')
-  const [oddsFormat, setOddsFormat] = useState<OddsFormat>(() => resolveInitialSportsEventOddsFormat())
-  const [showSpreadsAndTotals, setShowSpreadsAndTotals] = useState(() => resolveInitialShowSpreadsAndTotals())
-  const searchShellRef = useRef<HTMLDivElement | null>(null)
-  const searchInputRef = useRef<HTMLInputElement | null>(null)
-
-  return {
-    openCardId,
-    setOpenCardId,
-    isDetailsContentVisible,
-    setIsDetailsContentVisible,
-    activeDetailsTab,
-    setActiveDetailsTab,
-    selectedConditionByCardId,
-    setSelectedConditionByCardId,
-    tradeSelection,
-    setTradeSelection,
-    isSearchOpen,
-    setIsSearchOpen,
-    searchQuery,
-    setSearchQuery,
-    oddsFormat,
-    setOddsFormat,
-    showSpreadsAndTotals,
-    setShowSpreadsAndTotals,
-    searchShellRef,
-    searchInputRef,
-  }
-}
-
-function useCategoryResolver(categoryTitleBySlug: Record<string, string>) {
-  const normalizedCategoryTitleBySlug = useMemo(() => {
-    return Object.fromEntries(
-      Object.entries(categoryTitleBySlug).map(([slug, title]) => [slug.trim().toLowerCase(), title]),
-    )
-  }, [categoryTitleBySlug])
-  const resolveCardCategory = useCallback(
-    (card: SportsGamesCard) => resolveCardCategoryLabel(card, normalizedCategoryTitleBySlug),
-    [normalizedCategoryTitleBySlug],
-  )
-
-  return { resolveCardCategory }
-}
-
-function useOddsFormatAndSpreadsTotalsPersistence({
-  oddsFormat,
-  showSpreadsAndTotals,
-}: {
-  oddsFormat: OddsFormat
-  showSpreadsAndTotals: boolean
-}) {
-  useEffect(function persistOddsFormatAndSpreadsTotals() {
-    if (typeof window !== 'undefined') {
-      window.localStorage.setItem(SPORTS_EVENT_ODDS_FORMAT_STORAGE_KEY, oddsFormat)
-      window.localStorage.setItem(
-        SPORTS_GAMES_SHOW_SPREADS_TOTALS_STORAGE_KEY,
-        showSpreadsAndTotals ? '1' : '0',
-      )
-    }
-
-    return undefined
-  }, [oddsFormat, showSpreadsAndTotals])
-}
-
-function useResetMobileOrderPanelOnDeviceChange({
-  isMobile,
-  setIsMobileOrderPanelOpen,
-}: {
-  isMobile: boolean
-  setIsMobileOrderPanelOpen: (open: boolean) => void
-}) {
-  useEffect(function resetMobileOrderPanelOnDeviceChange() {
-    if (!isMobile) {
-      return undefined
-    }
-
-    // Avoid carrying over an open trade drawer while browsing cards on mobile.
-    setIsMobileOrderPanelOpen(false)
-    return undefined
-  }, [isMobile, setIsMobileOrderPanelOpen])
-}
-
-function useSportsGamesButtonOddsFormatter(oddsFormat: OddsFormat) {
-  const formatButtonOdds = useCallback((cents: number) => {
-    if (oddsFormat === 'price') {
-      return `${cents}¢`
-    }
-    return formatOddsFromCents(cents, oddsFormat)
-  }, [oddsFormat])
-
-  return { formatButtonOdds }
-}
-
-function useResolveDisplayButtonKey(showSpreadsAndTotals: boolean) {
-  const resolveDisplayButtonKey = useCallback((
-    card: SportsGamesCard,
-    preferredKey: string | null | undefined,
-  ) => {
-    const preferredButton = preferredKey
-      ? card.buttons.find(button => button.key === preferredKey) ?? null
-      : null
-    const visibleMarketTypes = new Set(resolveSportsGamesCardVisibleMarketTypes(card, showSpreadsAndTotals))
-    if (preferredButton && visibleMarketTypes.has(preferredButton.marketType)) {
-      return preferredButton.key
-    }
-
-    return card.buttons.find(button => visibleMarketTypes.has(button.marketType))?.key
-      ?? preferredButton?.key
-      ?? resolveDefaultConditionId(card)
-  }, [showSpreadsAndTotals])
-
-  return { resolveDisplayButtonKey }
-}
-
-function useVisiblePageCards({
-  cards,
-  isFeedPage,
-  isLivePage,
-  isSoonPage,
-  currentTimestampMs,
-}: {
-  cards: SportsGamesCard[]
-  isFeedPage: boolean
-  isLivePage: boolean
-  isSoonPage: boolean
-  currentTimestampMs: number
-}) {
-  const visibleCards = useMemo(() => {
-    if (isFeedPage) {
-      return cards
-    }
-
-    return cards.filter(card => !isSportsGamesCardResolved(card))
-  }, [cards, isFeedPage])
-
-  const pageCards = useMemo(() => {
-    if (isLivePage) {
-      return visibleCards.filter(card => isCardLiveNow(card, currentTimestampMs))
-    }
-
-    if (isSoonPage) {
-      return visibleCards.filter(card => isCardFuture(card, currentTimestampMs))
-    }
-
-    return visibleCards
-  }, [currentTimestampMs, isLivePage, isSoonPage, visibleCards])
-
-  return { visibleCards, pageCards }
-}
-
-function useWeekFilterState({
-  initialWeek,
-  isFeedPage,
-  visibleCards,
-  pageCards,
-}: {
-  initialWeek: number | null
-  isFeedPage: boolean
-  visibleCards: SportsGamesCard[]
-  pageCards: SportsGamesCard[]
-}) {
-  const weekOptions = useMemo(() => {
-    if (isFeedPage) {
-      return []
-    }
-
-    const weeks = Array.from(new Set(
-      visibleCards
-        .map(card => card.week)
-        .filter((week): week is number => Number.isFinite(week)),
-    ))
-
-    return weeks.sort((a, b) => a - b)
-  }, [isFeedPage, visibleCards])
-
-  const requestedWeekOption = useMemo(() => {
-    if (isFeedPage || initialWeek == null || !Number.isFinite(initialWeek)) {
-      return null
-    }
-    return String(initialWeek)
-  }, [initialWeek, isFeedPage])
-
-  const latestWeekOption = useMemo(
-    () => (weekOptions.length > 0 ? String(weekOptions.at(-1)) : 'all'),
-    [weekOptions],
-  )
-
-  const [selectedWeek, setSelectedWeek] = useState<string>(
-    requestedWeekOption
-    ?? latestWeekOption,
-  )
-
-  const effectiveSelectedWeek = useMemo(() => {
-    if (isFeedPage || weekOptions.length === 0) {
-      return 'all'
-    }
-
-    if (selectedWeek !== 'all') {
-      for (const week of weekOptions) {
-        if (String(week) === selectedWeek) {
-          return selectedWeek
-        }
-      }
-    }
-
-    if (requestedWeekOption != null) {
-      for (const week of weekOptions) {
-        if (String(week) === requestedWeekOption) {
-          return requestedWeekOption
-        }
-      }
-    }
-
-    return latestWeekOption
-  }, [isFeedPage, latestWeekOption, requestedWeekOption, selectedWeek, weekOptions])
-
-  const weekFilteredCards = useMemo(() => {
-    if (isFeedPage) {
-      return pageCards
-    }
-
-    if (effectiveSelectedWeek === 'all') {
-      return visibleCards
-    }
-
-    const week = Number(effectiveSelectedWeek)
-    return visibleCards.filter(card => card.week === week)
-  }, [effectiveSelectedWeek, isFeedPage, pageCards, visibleCards])
-
-  return {
-    weekOptions,
-    effectiveSelectedWeek,
-    setSelectedWeek,
-    weekFilteredCards,
-  }
-}
-
-function useSearchAutoFocus({
-  isSearchOpen,
-  searchInputRef,
-}: {
-  isSearchOpen: boolean
-  searchInputRef: React.RefObject<HTMLInputElement | null>
-}) {
-  useEffect(function focusSearchInputWhenSearchOpens() {
-    if (!isSearchOpen) {
-      return undefined
-    }
-    searchInputRef.current?.focus()
-    return undefined
-  }, [isSearchOpen, searchInputRef])
-}
-
-function useSearchOutsidePointerClose({
-  isSearchOpen,
-  searchQuery,
-  searchShellRef,
-  setIsSearchOpen,
-}: {
-  isSearchOpen: boolean
-  searchQuery: string
-  searchShellRef: React.RefObject<HTMLDivElement | null>
-  setIsSearchOpen: (open: boolean) => void
-}) {
-  useEffect(function closeSearchOnOutsidePointerDown() {
-    if (!isSearchOpen) {
-      return undefined
-    }
-
-    function handlePointerDown(event: PointerEvent) {
-      const target = event.target
-      if (!(target instanceof Node)) {
-        return
-      }
-      if (searchShellRef.current?.contains(target)) {
-        return
-      }
-      if (searchQuery.trim()) {
-        return
-      }
-      setIsSearchOpen(false)
-    }
-
-    window.addEventListener('pointerdown', handlePointerDown)
-    return function removeSearchPointerDownListener() {
-      window.removeEventListener('pointerdown', handlePointerDown)
-    }
-  }, [isSearchOpen, searchQuery, searchShellRef, setIsSearchOpen])
-}
-
-function useSportsSearchFilteredCards({
-  weekFilteredCards,
-  searchQuery,
-  resolveCardCategory,
-}: {
-  weekFilteredCards: SportsGamesCard[]
-  searchQuery: string
-  resolveCardCategory: (card: SportsGamesCard) => string
-}) {
-  const normalizedSearchQuery = useMemo(
-    () => normalizeComparableText(searchQuery),
-    [searchQuery],
-  )
-
-  const filteredCards = useMemo(() => {
-    if (!normalizedSearchQuery) {
-      return weekFilteredCards
-    }
-
-    return weekFilteredCards.filter((card) => {
-      const searchableText = [
-        card.title,
-        card.event.title,
-        card.event.slug,
-        resolveCardCategory(card),
-        ...(card.event.sports_tags ?? []),
-        ...card.teams.flatMap(team => [team.name, team.abbreviation]),
-      ]
-        .map(value => normalizeComparableText(value))
-        .join(' ')
-
-      return searchableText.includes(normalizedSearchQuery)
-    })
-  }, [normalizedSearchQuery, resolveCardCategory, weekFilteredCards])
-
-  return { normalizedSearchQuery, filteredCards }
-}
-
-function useCardButtonPriceMap(filteredCards: SportsGamesCard[]) {
-  const buttonTokenIds = useMemo(() => {
-    const tokenIds = new Set<string>()
-
-    filteredCards.forEach((card) => {
-      const marketsByConditionId = new Map(
-        card.detailMarkets.map(market => [market.condition_id, market] as const),
-      )
-
-      card.buttons.forEach((button) => {
-        const market = marketsByConditionId.get(button.conditionId)
-        const outcome = market?.outcomes.find(currentOutcome => currentOutcome.outcome_index === button.outcomeIndex)
-          ?? market?.outcomes[button.outcomeIndex]
-
-        if (outcome?.token_id) {
-          tokenIds.add(String(outcome.token_id))
-        }
-      })
-    })
-
-    return Array.from(tokenIds)
-  }, [filteredCards])
-  const { data: buttonOrderBookSummaries } = useOrderBookSummaries(buttonTokenIds)
-  const buttonPriceCentsByKey = useMemo(() => {
-    const priceByKey = new Map<string, number>()
-
-    filteredCards.forEach((card) => {
-      const marketsByConditionId = new Map(
-        card.detailMarkets.map(market => [market.condition_id, market] as const),
-      )
-
-      card.buttons.forEach((button) => {
-        const market = marketsByConditionId.get(button.conditionId) ?? null
-        const outcome = market?.outcomes.find(currentOutcome => currentOutcome.outcome_index === button.outcomeIndex)
-          ?? market?.outcomes[button.outcomeIndex]
-        const cents = resolveOutcomePriceCents(
-          market,
-          button.outcomeIndex === OUTCOME_INDEX.NO ? OUTCOME_INDEX.NO : OUTCOME_INDEX.YES,
-          {
-            orderBookSummaries: buttonOrderBookSummaries,
-            side: ORDER_SIDE.BUY,
-          },
-        )
-        const selectionCents = resolveOutcomeSelectionPriceCents(market, outcome, {
-          orderBookSummaries: buttonOrderBookSummaries,
-          side: ORDER_SIDE.BUY,
-          fallbackIsNoOutcome: button.fallbackIsNoOutcome,
-        })
-        priceByKey.set(`${card.id}:${button.key}`, selectionCents ?? cents ?? button.cents)
-      })
-    })
-
-    return priceByKey
-  }, [buttonOrderBookSummaries, filteredCards])
-
-  return { buttonPriceCentsByKey }
-}
-
-function useEffectiveOpenAndTradeSelection({
-  openCardId,
-  filteredCards,
-  tradeSelection,
-  selectedConditionByCardId,
-  showSpreadsAndTotals,
-  resolveDisplayButtonKey,
-}: {
-  openCardId: string | null
-  filteredCards: SportsGamesCard[]
-  tradeSelection: SportsTradeSelection
-  selectedConditionByCardId: Record<string, string>
-  showSpreadsAndTotals: boolean
-  resolveDisplayButtonKey: (card: SportsGamesCard, preferredKey: string | null | undefined) => string | null
-}) {
-  const effectiveOpenCardId = useMemo(() => {
-    if (!openCardId) {
-      return null
-    }
-
-    return filteredCards.some(card => card.id === openCardId) ? openCardId : null
-  }, [filteredCards, openCardId])
-
-  const effectiveTradeSelection = useMemo<SportsTradeSelection>(() => {
-    if (filteredCards.length === 0) {
-      return { cardId: null, buttonKey: null }
-    }
-
-    const currentCard = tradeSelection.cardId
-      ? filteredCards.find(card => card.id === tradeSelection.cardId) ?? null
-      : null
-
-    if (currentCard) {
-      const currentButton = tradeSelection.buttonKey
-        ? currentCard.buttons.find(button => button.key === tradeSelection.buttonKey) ?? null
-        : null
-      const currentButtonExists = Boolean(
-        currentButton
-        && (showSpreadsAndTotals || currentButton.marketType === 'moneyline'),
-      )
-      if (currentButtonExists) {
-        return tradeSelection
-      }
-
-      const preferredButtonKey = resolveDisplayButtonKey(
-        currentCard,
-        selectedConditionByCardId[currentCard.id] ?? resolveDefaultConditionId(currentCard),
-      )
-      const fallbackButtonKey = resolveSelectedButton(currentCard, preferredButtonKey)?.key ?? null
-      return {
-        cardId: currentCard.id,
-        buttonKey: fallbackButtonKey,
-      }
-    }
-
-    const firstCard = filteredCards[0]
-    const preferredButtonKey = resolveDisplayButtonKey(
-      firstCard,
-      selectedConditionByCardId[firstCard.id] ?? resolveDefaultConditionId(firstCard),
-    )
-    const firstButtonKey = resolveSelectedButton(firstCard, preferredButtonKey)?.key ?? null
-    return {
-      cardId: firstCard.id,
-      buttonKey: firstButtonKey,
-    }
-  }, [filteredCards, resolveDisplayButtonKey, selectedConditionByCardId, showSpreadsAndTotals, tradeSelection])
-
-  return { effectiveOpenCardId, effectiveTradeSelection }
-}
-
-function useLocaleDateTimeFormatters(locale: string) {
-  const dateLabelFormatter = useMemo(
-    () => new Intl.DateTimeFormat(locale, {
-      weekday: 'short',
-      month: 'long',
-      day: 'numeric',
-    }),
-    [locale],
-  )
-
-  const timeLabelFormatter = useMemo(
-    () => new Intl.DateTimeFormat(locale, {
-      hour: 'numeric',
-      minute: '2-digit',
-    }),
-    [locale],
-  )
-
-  return { dateLabelFormatter, timeLabelFormatter }
-}
-
-function useCardGroupings({
-  filteredCards,
-  dateLabelFormatter,
-  resolveCardCategory,
-  currentTimestampMs,
-}: {
-  filteredCards: SportsGamesCard[]
-  dateLabelFormatter: Intl.DateTimeFormat
-  resolveCardCategory: (card: SportsGamesCard) => string
-  currentTimestampMs: number
-}) {
-  const groupedCards = useMemo(() => {
-    const grouped = new Map<string, { key: string, label: string, sortValue: number, cards: SportsGamesCard[] }>()
-
-    for (const card of filteredCards) {
-      const date = card.startTime ? new Date(card.startTime) : null
-      const isValidDate = Boolean(date && !Number.isNaN(date.getTime()))
-      const groupKey = isValidDate ? toDateGroupKey(date as Date) : 'tbd'
-      const label = isValidDate ? dateLabelFormatter.format(date as Date) : 'Date TBD'
-      const sortValue = isValidDate ? (date as Date).getTime() : Number.POSITIVE_INFINITY
-
-      const existing = grouped.get(groupKey)
-      if (existing) {
-        existing.cards.push(card)
-        continue
-      }
-
-      grouped.set(groupKey, {
-        key: groupKey,
-        label,
-        sortValue,
-        cards: [card],
-      })
-    }
-
-    return Array.from(grouped.values()).sort((a, b) => a.sortValue - b.sortValue)
-  }, [dateLabelFormatter, filteredCards])
-
-  const liveCards = useMemo(
-    () => filteredCards.filter(card => isCardLiveNow(card, currentTimestampMs)),
-    [currentTimestampMs, filteredCards],
-  )
-
-  const liveCardsByCategory = useMemo(() => {
-    const grouped = new Map<string, { key: string, label: string, cards: SportsGamesCard[] }>()
-
-    for (const card of liveCards) {
-      const label = resolveCardCategory(card)
-      const key = normalizeComparableText(label) || card.id
-      const existing = grouped.get(key)
-      if (existing) {
-        existing.cards.push(card)
-        continue
-      }
-      grouped.set(key, {
-        key,
-        label,
-        cards: [card],
-      })
-    }
-
-    return Array.from(grouped.values())
-      .sort((left, right) => left.label.localeCompare(right.label))
-  }, [liveCards, resolveCardCategory])
-
-  const sortedFutureCards = useMemo(() => {
-    const future = filteredCards.filter(card => isCardFuture(card, currentTimestampMs))
-
-    return [...future].sort((left, right) => {
-      const leftStart = resolveCardStartTimestamp(left)
-      const rightStart = resolveCardStartTimestamp(right)
-      const leftHasStart = Number.isFinite(leftStart)
-      const rightHasStart = Number.isFinite(rightStart)
-
-      if (leftHasStart && rightHasStart) {
-        return leftStart - rightStart
-      }
-      if (leftHasStart) {
-        return -1
-      }
-      if (rightHasStart) {
-        return 1
-      }
-
-      return left.id.localeCompare(right.id)
-    })
-  }, [currentTimestampMs, filteredCards])
-
-  const startingSoonGroupsByDate = useMemo(() => {
-    const groupedByDate = new Map<
-      string,
-      {
-        key: string
-        label: string
-        sortValue: number
-        categories: Map<string, { key: string, label: string, cards: SportsGamesCard[] }>
-      }
-    >()
-
-    for (const card of sortedFutureCards) {
-      const startMs = resolveCardStartTimestamp(card)
-      const date = Number.isFinite(startMs) ? new Date(startMs) : null
-      const isValidDate = Boolean(date && !Number.isNaN(date.getTime()))
-      const dateKey = isValidDate ? toDateGroupKey(date as Date) : 'tbd'
-      const dateLabel = isValidDate ? dateLabelFormatter.format(date as Date) : 'Date TBD'
-      const sortValue = isValidDate ? (date as Date).getTime() : Number.POSITIVE_INFINITY
-
-      const categoryLabel = resolveCardCategory(card)
-      const categoryKey = normalizeComparableText(categoryLabel) || card.id
-
-      const existingDateGroup = groupedByDate.get(dateKey)
-      if (existingDateGroup) {
-        const existingCategory = existingDateGroup.categories.get(categoryKey)
-        if (existingCategory) {
-          existingCategory.cards.push(card)
-        }
-        else {
-          existingDateGroup.categories.set(categoryKey, {
-            key: categoryKey,
-            label: categoryLabel,
-            cards: [card],
-          })
-        }
-        continue
-      }
-
-      groupedByDate.set(dateKey, {
-        key: dateKey,
-        label: dateLabel,
-        sortValue,
-        categories: new Map([
-          [categoryKey, {
-            key: categoryKey,
-            label: categoryLabel,
-            cards: [card],
-          }],
-        ]),
-      })
-    }
-
-    return Array.from(groupedByDate.values())
-      .sort((left, right) => left.sortValue - right.sortValue)
-      .map(group => ({
-        key: group.key,
-        label: group.label,
-        sortValue: group.sortValue,
-        categories: Array.from(group.categories.values())
-          .sort((left, right) => left.label.localeCompare(right.label)),
-      }))
-  }, [dateLabelFormatter, resolveCardCategory, sortedFutureCards])
-
-  return {
-    groupedCards,
-    liveCardsByCategory,
-    startingSoonGroupsByDate,
-  }
-}
-
-function useSportsActiveTradeContext({
-  effectiveOpenCardId,
-  effectiveTradeSelection,
-  filteredCards,
-  resolveDisplayButtonKey,
-  selectedConditionByCardId,
-  orderMarketConditionId,
-  orderOutcomeIndex,
-}: {
-  effectiveOpenCardId: string | null
-  effectiveTradeSelection: SportsTradeSelection
-  filteredCards: SportsGamesCard[]
-  resolveDisplayButtonKey: (card: SportsGamesCard, preferredKey: string | null | undefined) => string | null
-  selectedConditionByCardId: Record<string, string>
-  orderMarketConditionId: string | null
-  orderOutcomeIndex: number | null
-}) {
-  const activeTradeContext = useMemo<SportsActiveTradeContext | null>(() => {
-    if (filteredCards.length === 0) {
-      return null
-    }
-
-    const selectedCardFromTrade = effectiveTradeSelection.cardId
-      ? filteredCards.find(card => card.id === effectiveTradeSelection.cardId) ?? null
-      : null
-    const selectedCardFromOpen = effectiveOpenCardId
-      ? filteredCards.find(card => card.id === effectiveOpenCardId) ?? null
-      : null
-    const card = selectedCardFromTrade ?? selectedCardFromOpen ?? filteredCards[0] ?? null
-    if (!card) {
-      return null
-    }
-
-    const selectedButtonKey = resolveDisplayButtonKey(card, (
-      effectiveTradeSelection.cardId === card.id
-        ? effectiveTradeSelection.buttonKey
-        : null
-    ) ?? selectedConditionByCardId[card.id] ?? resolveDefaultConditionId(card))
-
-    const button = resolveSelectedButton(card, selectedButtonKey)
-    if (!button) {
-      return null
-    }
-
-    const market = resolveSelectedMarket(card, button.key)
-    if (!market) {
-      return null
-    }
-
-    const outcome = resolveSelectedOutcome(market, button)
-    if (!outcome) {
-      return null
-    }
-
-    return {
-      card,
-      button,
-      market,
-      outcome,
-    }
-  }, [effectiveOpenCardId, effectiveTradeSelection.buttonKey, effectiveTradeSelection.cardId, filteredCards, resolveDisplayButtonKey, selectedConditionByCardId])
-
-  const activeTradePrimaryOutcomeIndex = useMemo(() => {
-    if (!activeTradeContext || activeTradeContext.button.marketType !== 'spread') {
-      return null
-    }
-
-    return resolveStableSpreadPrimaryOutcomeIndex(
-      activeTradeContext.card,
-      activeTradeContext.button.conditionId,
-    )
-  }, [activeTradeContext])
-
-  const activeTradeHeaderContext = useMemo<SportsActiveTradeContext | null>(() => {
-    if (!activeTradeContext) {
-      return null
-    }
-
-    if (!orderMarketConditionId || orderMarketConditionId !== activeTradeContext.market.condition_id) {
-      return activeTradeContext
-    }
-
-    if (orderOutcomeIndex == null) {
-      return activeTradeContext
-    }
-
-    const matchedOutcome = activeTradeContext.market.outcomes.find(
-      outcome => outcome.outcome_index === orderOutcomeIndex,
-    ) ?? activeTradeContext.outcome
-
-    const matchedButton = activeTradeContext.card.buttons.find(
-      button => (
-        button.conditionId === activeTradeContext.market.condition_id
-        && button.outcomeIndex === orderOutcomeIndex
-      ),
-    ) ?? activeTradeContext.button
-
-    return {
-      ...activeTradeContext,
-      button: matchedButton,
-      outcome: matchedOutcome,
-    }
-  }, [activeTradeContext, orderMarketConditionId, orderOutcomeIndex])
-  const orderPanelOutcomeLabelOverrides = useMemo(
-    () => activeTradeContext
-      ? resolveOrderPanelOutcomeLabelOverrides(
-          activeTradeHeaderContext?.card ?? activeTradeContext.card,
-          activeTradeHeaderContext?.market ?? activeTradeContext.market,
-        )
-      : {},
-    [activeTradeContext, activeTradeHeaderContext],
-  )
-  const orderPanelOutcomeAccentOverrides = useMemo(
-    () => activeTradeContext
-      ? resolveOrderPanelOutcomeAccentOverrides(
-          activeTradeHeaderContext?.card ?? activeTradeContext.card,
-          activeTradeHeaderContext?.market ?? activeTradeContext.market,
-        )
-      : {},
-    [activeTradeContext, activeTradeHeaderContext],
-  )
-
-  return {
-    activeTradeContext,
-    activeTradePrimaryOutcomeIndex,
-    activeTradeHeaderContext,
-    orderPanelOutcomeLabelOverrides,
-    orderPanelOutcomeAccentOverrides,
-  }
-}
-
-function useSportsOrderStoreSync({
-  activeTradeContext,
-  setOrderEvent,
-  setOrderMarket,
-  setOrderOutcome,
-  setOrderSide,
-}: {
-  activeTradeContext: SportsActiveTradeContext | null
-  setOrderEvent: (event: Event) => void
-  setOrderMarket: (market: Market) => void
-  setOrderOutcome: (outcome: Outcome) => void
-  setOrderSide: (side: typeof ORDER_SIDE.BUY | typeof ORDER_SIDE.SELL) => void
-}) {
-  useEffect(function syncSportsOrderStoreFromActiveTrade() {
-    if (!activeTradeContext) {
-      return undefined
-    }
-
-    const {
-      event: currentOrderEvent,
-      market: currentOrderMarket,
-      outcome: currentOrderOutcome,
-    } = useOrder.getState()
-
-    const isSameSelection = (
-      currentOrderEvent?.id === activeTradeContext.card.event.id
-      && currentOrderMarket?.condition_id === activeTradeContext.market.condition_id
-      && currentOrderOutcome?.outcome_index === activeTradeContext.outcome.outcome_index
-    )
-
-    if (currentOrderEvent !== activeTradeContext.card.event) {
-      setOrderEvent(activeTradeContext.card.event)
-    }
-
-    if (currentOrderMarket !== activeTradeContext.market) {
-      setOrderMarket(activeTradeContext.market)
-    }
-
-    if (currentOrderOutcome !== activeTradeContext.outcome) {
-      setOrderOutcome(activeTradeContext.outcome)
-    }
-
-    if (!isSameSelection) {
-      setOrderSide(ORDER_SIDE.BUY)
-    }
-    return undefined
-  }, [
-    activeTradeContext,
-    setOrderEvent,
-    setOrderMarket,
-    setOrderOutcome,
-    setOrderSide,
-  ])
-}
+import { headerIconButtonClass, MARKET_COLUMN_BY_KEY } from './_sports-games-center/sports-games-center-constants'
+import {
+  groupButtonsByMarketType,
+  isCardLiveNow,
+  parseSportsScore,
+  resolveActiveMarketType,
+  resolveButtonDepthStyle,
+  resolveButtonOverlayStyle,
+  resolveButtonStyle,
+  resolveDefaultConditionId,
+  resolveMoneylineButtonGridClass,
+  resolveSelectedButton,
+} from './_sports-games-center/sports-games-center-utils'
+import SportsGameDetailsPanel from './_sports-games-center/SportsGameDetailsPanel'
+import SportsOrderPanelMarketInfo from './_sports-games-center/SportsOrderPanelMarketInfo'
+import {
+  useCardButtonPriceMap,
+  useCardGroupings,
+  useCategoryResolver,
+  useEffectiveOpenAndTradeSelection,
+  useLocaleDateTimeFormatters,
+  useOddsFormatAndSpreadsTotalsPersistence,
+  useResetMobileOrderPanelOnDeviceChange,
+  useResolveDisplayButtonKey,
+  useSearchAutoFocus,
+  useSearchOutsidePointerClose,
+  useSportsActiveTradeContext,
+  useSportsGamesButtonOddsFormatter,
+  useSportsGamesCenterShellState,
+  useSportsOrderStoreSync,
+  useSportsSearchFilteredCards,
+  useVisiblePageCards,
+  useWeekFilterState,
+} from './_sports-games-center/useSportsGamesCenter'
+
+// Re-export types and values consumed by SportsEventCenter and other external files
+export type { SportsGameGraphVariant, SportsGamesMarketType, SportsLinePickerOption } from './_sports-games-center/sports-games-center-types'
+export {
+  buildLinePickerOptions,
+  groupButtonsByMarketType,
+  resolveButtonDepthStyle,
+  resolveButtonOverlayStyle,
+  resolveButtonStyle,
+  resolveDefaultConditionId,
+  resolveOrderPanelOutcomeAccentOverrides,
+  resolveOrderPanelOutcomeLabelOverrides,
+  resolvePreferredLinePickerButton,
+  resolveSelectedButton,
+  resolveSelectedMarket,
+  resolveSelectedOutcome,
+  resolveStableSpreadPrimaryOutcomeIndex,
+} from './_sports-games-center/sports-games-center-utils'
+export { default as SportsGameDetailsPanel } from './_sports-games-center/SportsGameDetailsPanel'
+export { default as SportsGameGraph } from './_sports-games-center/SportsGameGraph'
+export { default as SportsOrderPanelMarketInfo } from './_sports-games-center/SportsOrderPanelMarketInfo'
 
 export default function SportsGamesCenter({
   cards,
@@ -5706,7 +757,7 @@ export default function SportsGamesCenter({
                   let renderedButtons = columnButtons
 
                   if (column.key !== 'moneyline') {
-                    const buttonsByConditionId = new Map<string, SportsGamesButton[]>()
+                    const buttonsByConditionId = new Map<string, typeof columnButtons>()
                     for (const button of columnButtons) {
                       const existing = buttonsByConditionId.get(button.conditionId)
                       if (existing) {
@@ -5728,7 +779,7 @@ export default function SportsGamesCenter({
                     renderedButtons = selectedButtons
 
                     if (column.key === 'spread') {
-                      const spreadOrder: Record<SportsGamesButton['tone'], number> = {
+                      const spreadOrder: Record<string, number> = {
                         team1: 0,
                         team2: 1,
                         draw: 2,
